% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_tags.R
\name{add_field}
\alias{add_field}
\alias{add_field_seq}
\title{Add a field to neuron objects (inc. in a neuronlist)}
\usage{
add_field(x, entry, field = "bodyid", ...)

add_field_seq(x, entries, field = "bodyid", ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by
\code{neuprintr::neuprint_read_neurons} or possibly
\code{catmaid::read.neurons.catmaid}.}

\item{entry}{item to add.}

\item{field}{name of new field, or field to overwrite.}

\item{...}{Additional arguments passed to methods or eventually to
\code{nat::\link{nlapply}}}

\item{entries}{a vector/list of entries, the same length as \code{x}.
Each neuron in \code{x} will have the entry in entries as the same index, added to it
as \code{field}.}
}
\value{
a \code{neuron} or \code{neuronlist}
}
\description{
Add an item to the the list object that comprises a neuron
}
\seealso{
\code{\link{add_Label}}
}
