% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{classed.ids}
\alias{classed.ids}
\alias{rn.ids}
\alias{orn.ids}
\alias{hrn.ids}
\alias{pn.ids}
\alias{upn.ids}
\alias{mpn.ids}
\alias{vppn.ids}
\alias{dan.ids}
\alias{mbon.ids}
\title{Bodyids for neuron classes}
\format{An object of class \code{character} of length 2545.}
\usage{
rn.ids

orn.ids

hrn.ids

pn.ids

upn.ids

mpn.ids

vppn.ids

dan.ids

mbon.ids
}
\description{
Bodyids for neuron classes. Hard saved equivalent for \code{\link{class2ids}}.
However, you should use \code{\link{class2ids}} for the most up-to-date information.
}
\details{
neuPrint bodyids for neurons that the FlyConnectome groups at the University of Cambridge, dept. Zoology,
have determined to fall into different classes.
}
\examples{
\donttest{
head(rn.ids)
}
}
\seealso{
\code{\link{class2ids}}
}
\keyword{datasets}
