% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_cable.R
\name{axonic_points}
\alias{axonic_points}
\alias{dendritic_points}
\alias{mixed_points}
\alias{primary_dendrite_points}
\alias{axonic_points.neuron}
\alias{dendritic_points.neuron}
\alias{mixed_points.neuron}
\alias{primary_dendrite_points.neuron}
\alias{dendritic_points.neuronlist}
\alias{axonic_points.neuronlist}
\alias{mixed_points.neuronlist}
\alias{primary_dendrite_points.neuronlist}
\alias{axonal_endings}
\alias{dendritic_endings}
\alias{axonic_endings}
\alias{primary_dendrite_endings}
\alias{axonic_cable}
\alias{dendritic_cable}
\alias{arbour_cable}
\alias{unsure_cable}
\alias{primary_dendrite_cable}
\alias{axonic_cable.neuron}
\alias{axonic_cable.catmaidneuron}
\alias{dendritic_cable.neuron}
\alias{dendritic_cable.catmaidneuron}
\alias{arbour_cable.neuron}
\alias{arbour_cable.catmaidneuron}
\alias{unsure_cable.neuron}
\alias{unsure_cable.catmaidneuron}
\alias{primary_dendrite_cable.neuron}
\alias{primary_dendrite_cable.catmaidneuron}
\alias{axonic_cable.neuronlist}
\alias{dendritic_cable.neuronlist}
\alias{arbour_cable.neuronlist}
\alias{unsure_cable.neuronlist}
\alias{primary_dendrite_cable.neuronlist}
\title{Extract axonic/dendritic points/cable from a neuron/neuronlist}
\usage{
axonic_points(x, ...)

dendritic_points(x, ...)

mixed_points(x, ...)

primary_dendrite_points(x, ...)

\method{axonic_points}{neuron}(x)

\method{dendritic_points}{neuron}(x)

\method{mixed_points}{neuron}(x)

\method{primary_dendrite_points}{neuron}(x)

\method{dendritic_points}{neuronlist}(x, ...)

\method{axonic_points}{neuronlist}(x, ...)

\method{mixed_points}{neuronlist}(x, ...)

\method{primary_dendrite_points}{neuronlist}(x, ...)

axonal_endings(x)

dendritic_endings(x)

axonic_endings(x)

primary_dendrite_endings(x)

axonic_cable(x, ...)

dendritic_cable(x, ...)

arbour_cable(x, ...)

unsure_cable(x, ...)

primary_dendrite_cable(x, ...)

\method{axonic_cable}{neuron}(x, mixed = FALSE, ...)

\method{axonic_cable}{catmaidneuron}(x, mixed = FALSE, ...)

\method{dendritic_cable}{neuron}(x, mixed = FALSE, ...)

\method{dendritic_cable}{catmaidneuron}(x, mixed = FALSE, ...)

\method{arbour_cable}{neuron}(x, mixed = FALSE, ...)

\method{arbour_cable}{catmaidneuron}(x, mixed = FALSE, ...)

\method{unsure_cable}{neuron}(x, mixed = FALSE, ...)

\method{unsure_cable}{catmaidneuron}(x, mixed = FALSE, ...)

\method{primary_dendrite_cable}{neuron}(x, ...)

\method{primary_dendrite_cable}{catmaidneuron}(x, ...)

\method{axonic_cable}{neuronlist}(x, mixed = FALSE, ...)

\method{dendritic_cable}{neuronlist}(x, mixed = FALSE, ...)

\method{arbour_cable}{neuronlist}(x, mixed = FALSE, ...)

\method{unsure_cable}{neuronlist}(x, ...)

\method{primary_dendrite_cable}{neuronlist}(x, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object that has its axons/dendrites labelled in swc format in its neuron$d dataframes}

\item{...}{additional arguments passed to methods}

\item{mixed}{whether or not to include points assigned as uncertain or mixed polarity cable}
}
\value{
a matrix of 3D points
}
\description{
Extract axonic/dendritic/mixed/primary dendrite points/endpoints/cable from a neuron/neuronlist object
}
