% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_meta.R
\name{fafb_hemibrain_annotate}
\alias{fafb_hemibrain_annotate}
\alias{fafb_set_hemilineage}
\title{Set annotations for FAFB neurons in CATMAID based on hemibrain results}
\usage{
fafb_hemibrain_annotate(x, ...)

fafb_set_hemilineage(
  find,
  putative = TRUE,
  ItoLee_Hemilineage,
  delete.find = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector of FAFB skeleton IDs (skids) that can be read with \code{catmaid::catmaid_skids()}}

\item{...}{arguments passed to \code{neuprintr::neuprint_get_meta} and \code{catmaid::catmaid_login}.}

\item{find}{an annotation/search term/vector of skids passed to \code{catmaid::catmaid_skids}.}

\item{ItoLee_Hemilineage}{character, the correct K. Ito / T. Lee hemilineage. Must be an entry in \code{\link{hemibrain_hemilineages}}. If set to 'wipe' all lineage related annotations will be removed.}

\item{delete.find}{logical, is \code{TRUE} then if \code{find} is an annotation, it will be wiped from the neuron after this function is used (if you have permission to remove it)}
}
\value{
annotations set on CATMAID neurons in specified CATMAID instance, see your \code{catmaid::catmaid_connection()}
}
\description{
Set annotations for FAFB neurons in CATMAID based on matches made to hemibrain neurons.
E.g. transfer information on matches and cell body fibers, and also update lineage related information.
}
\examples{
\donttest{
\dontrun{

# Transfer information on the VA6 uPN
fafb_hemibrain_annotate("16")

}}
}
\seealso{
\code{\link{hemibrain_get_meta}}
}
