% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_googledrive.R
\name{flycircuit_neurons}
\alias{flycircuit_neurons}
\alias{lm_lhns}
\title{Read flycircuit neurons from hemibrainr Google Drive}
\usage{
flycircuit_neurons(
  local = FALSE,
  folder = "light_level/flycircuit",
  cable = c("all", "primary.neurites", "arbour"),
  data = c("neuronlist", "nblast", "dps"),
  brainspace = c("JRCFIB2018F", "JRCFIB2018Fraw", "FCWB")
)

lm_lhns(
  local = FALSE,
  folder = "light_level/lhns",
  data = c("neuronlist", "nblast", "dps"),
  cable = c("lhns", "lhins", "lines"),
  brainspace = c("JRCFIB2018F", "JRCFIB2018Fraw", "FCWB")
)
}
\arguments{
\item{local}{\code{FALSE} or path. By default (\code{FALSE}) data is read from \code{options()$Drive_hemibrain_data}),
but the user can specify an alternative path.}

\item{folder}{A subfolder on the hemibrain team drive or your local data folder
containing the data object to read.}

\item{cable}{the type of cable we want to read. \code{"all"} indicates full neurons.}

\item{data}{the type of data to read, i.e. neurons, an NBLAST matrix or a
\code{\link{dotprops}} object.}

\item{brainspace}{A template brain space for neurons loaded by
\code{lm_lhns}. Defaults to \code{JRCFIB2018F}.}
}
\value{
a \code{nat::neuronlistfh} object for 'light-level' neurons in the given brainspace.
}
\description{
Read \href{http://www.flycircuit.tw/}{flycircuit} neurons from the \code{hemibrainr} google drive.
The google drive must be mounted with Google Filestream or rclone.
See \code{\link{hemibrainr_set_drive}}.
}
\examples{
\donttest{
\dontrun{

# All flycircuit neurons
fw = flycircuit_neurons()

# All LHNs from Frechter et al. 2019, eLife
lhns = lm_lhns()

}}
}
\seealso{
\code{\link{hemibrain_splitpoints}},
  \code{\link{hemibrain_flow_centrality}},
  \code{\link{hemibrain_metrics}},
  \code{\link{hemibrainr_googledrive_data}}
}
