% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire_tracing.R
\name{flywire_annotations_for_synapses}
\alias{flywire_annotations_for_synapses}
\title{Generate a CSV of neuron synapses to import to a flywire annotation layer}
\usage{
flywire_annotations_for_synapses(
  fw.ids,
  direction = c("inputs", "oputputs", "both"),
  partners = NULL,
  db = flywire_neurons(WithConnectors = TRUE),
  keep.dist.nm = 500,
  cleft_scores.thresh = 75,
  sample = 250,
  write.csv = TRUE,
  csv.path = getwd(),
  ...
)
}
\arguments{
\item{fw.ids}{character vector, a vector of valid flywire IDs}

\item{direction}{whether to get presynapses, postsynapses or both.}

\item{partners}{character vector else not used i \code{NULL}. A vector of
valid flywire IDs for postsynaptic neurons to keep. Synapses to other
postsynaptic targets are filtered out.}

\item{db}{a \code{neuronlist} of flywire neurons with synapses attached}

\item{keep.dist.nm}{numeric, minimum distance in nm that one synapse can be
from another. One random synapse is chosen from groups of proximal
synapses. To deactivate, enter \code{NULL}.}

\item{cleft_scores.thresh}{numeric, cleft_score threshold for synapse
inclusion}

\item{sample}{numeric, the number of synapses to choose from both the
dendrite and axon of each neuron in \code{fw.ids}}

\item{write.csv}{logical, whether or not to write a \code{.csv} output file,
ready for import into flywire. One for each neuron, named by
\code{csv.path}.}

\item{...}{further arguments passed to \code{fafbseg::flywire_partners}, when \code{db} is \code{NULL}.}

\item{xyz}{a \code{data.frame} of xyz coordinates in flywire voxel space
to write as a flyiwre annotation .csv file. Can contains other columns.}

\item{description}{the column in \code{xyz} to use for the Descritpion field of the final \code{.csv}.}
}
\description{
Generate a CSV of neuron synapses to import to a flywire annotation layer
}
