% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire_googledrive.R
\name{flywire_neurons}
\alias{flywire_neurons}
\title{Download a large set of well-traced skeletonised neurons from FlyWire}
\usage{
flywire_neurons(
  local = FALSE,
  brain = c("FlyWire", "JRCFIB2018F", "JRCFIB2018F", "FAFB", "JFRC2", "JRC2018F",
    "FCWB"),
  mirror = FALSE
)
}
\arguments{
\item{brain}{the brainspace in which hemibrain neurons have been registered. Defaults to raw voxel space for the FlyWire project.}

\item{...}{Additional arguments passed to \code{\link{nlapply}} when iterating
over neurons to download.}
}
\description{
Get a large number of skeletonised neurons from FlyWire. The neuronlist is saved as a
 \code{nat::neuronlistfh} object so certain neurons may be read from it
 without loading the entire, large neuronlist into memory. You will need access to the hemibrain Google Team Drive and
 have it mounted with Google filestream.
}
\examples{
\donttest{
\dontrun{
fw.neurons = flywire_neurons()
}}
}
\seealso{
\code{\link{hemibrain_read_neurons}}
}
