% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire_synapses.R
\name{flywire_synapse_simplify}
\alias{flywire_synapse_simplify}
\title{Collapse Buhmann predicted flywire synapses into smaller units}
\usage{
flywire_synapse_simplify(
  x,
  method = c("cleft_scores", "scores", "mean"),
  collapse = TRUE,
  fast = TRUE
)
}
\arguments{
\item{x}{a  \code{data.frame} of synapses for FlyWire neurons, as generated by, for example, \code{fafbseg::flywire_partners}.}

\item{method}{when synapses are collapsed, they are either represented by the predicted synapse amongst them with the best score, cleft_score or a mean of
the meat data for each of these predicted synapses.}

\item{collapse}{logical, whether or not to collapse synapses in the same cluster to a single synapse. If \code{FALSE} then cluster groupings are given in the returned \code{data.frame}.}

\item{fast}{logical, whether or not to use \code{fastcluster::hclust} rather than \code{stats::hclust} to cluster synapses.}
}
\value{
a \code{data.frame} of collapsed synapses.
}
\description{
Many predicted Buhmann synapses may be located at one real synapse. Collapse Buhmann predicted flywire synapses into smaller units using a hierarchical clustering
based on the Eucliean distances between predicted 'presynapses'.
}
\examples{
\donttest{
\dontrun{
syns=fafbseg::flywire_partners("720575940621039145")
syns.simp = flywire_synapse_simplify(syns)
}}
}
\seealso{
\code{\link{flywire_googledrive_data}}
}
