% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron_morphology.R
\name{geodesic_distance}
\alias{geodesic_distance}
\alias{geodesic_distance.neuron}
\alias{geodesic_distance.neuronlist}
\title{Calculate the geodesic distance through an arbour between synapses.}
\usage{
geodesic_distance(x, graph.distance = TRUE, Label = 1, ...)

\method{geodesic_distance}{neuron}(x, graph.distance = TRUE, Label = 1, ...)

\method{geodesic_distance}{neuronlist}(x, graph.distance = TRUE, Label = 1, ...)
}
\arguments{
\item{x}{a \code{neuron} or \code{neuronlist} object.}

\item{graph.distance}{logical, whether to use geodesic distance in terms of node number (\code{TRUE}) or real unit distance.}

\item{Label}{defaults to 1, so distances to soma are chosen. If 7, then distance from primary neurite's branchpoint.}

\item{...}{methods passed to \code{nat::nlapply}}
}
\description{
Calculate the geodesic distance through an arbour between synapses.
}
