% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{glomerulus}
\alias{glomerulus}
\title{Find the antennal lobe glomerulus for hemibrain body ids}
\usage{
glomerulus(bodyids, exclude.multi = FALSE)
}
\arguments{
\item{bodyids}{Vector of bodyids or a search string passed to
\code{\link{neuprint_ids}}}

\item{exclude.multi}{When \code{TRUE} multiglomerular neurons will be
returned as \code{NA}.}
}
\value{
A character vector of glomeruli named with the \code{bodyids}
}
\description{
Find the antennal lobe glomerulus for hemibrain body ids
}
\examples{
\donttest{
table(glomerulus(class2ids('PN')), useNA='ifany')
table(glomerulus(class2ids('PN'), exclude.multi = TRUE), useNA='ifany')
}
}
