% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{grouped_adjacency}
\alias{grouped_adjacency}
\title{Flexible grouping of rows/columns of a raw adjacency matrix}
\usage{
grouped_adjacency(
  inputids = NULL,
  outputids = NULL,
  ingroup = c("type", "name", "cellBodyFiber"),
  outgroup = c("type", "name", "cellBodyFiber"),
  threshold = 0,
  scale = c("none", "col", "row"),
  ...
)
}
\arguments{
\item{inputids}{Either the bodyids of the input neurons OR an adjacency
matrix}

\item{outputids}{bodyids of the output neurons. Not required if
\code{inputids} is an adjacency matrix.}

\item{ingroup, outgroup}{grouping variables used to coalesce related neurons}

\item{threshold}{Remove (groups of) neurons that make fewer connections than
this.}

\item{scale}{Whether to scale the rows or columns so that they sum to 1}

\item{...}{Additional arguments passed to
\code{\link{neuprint_get_adjacency_matrix}}}
}
\value{
named numeric matrix
}
\description{
Use this to create a summary adjacency matrix in which input and/or output
neurons are coalesced into a smaller number of rows/columns. This can be used
as a preprocessing step to present a simple representation of the
connectivity between a larger number of cells in a heatmap.
}
\examples{
\donttest{
pnkc=grouped_adjacency(class2ids("PN"), 'KC')
heatmap(pnkc)
}
}
