% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain.surf}
\alias{hemibrain.surf}
\title{The hemibrain mesh}
\format{An object of class \code{hxsurf} (inherits from \code{list}) of length 4.}
\source{
\url{https://neuprint.janelia.org/}
}
\usage{
hemibrain.surf
}
\description{
A 3D mesh of the hemibrain volume as a
\code{\link[nat]{hxsurf}} object. Created by
Philipp Schlegel from the individual ROIs in
FlyEM's hemibrain project.
}
\examples{
\donttest{
rgl::wire3d(hemibrain)
rgl::shade3d(hemibrain, alpha=.3)
rgl::shade3d(hemibrain, alpha=.3, col='grey')
}
}
\seealso{
\code{\link{hemibrain_roi_meshes}}
}
\keyword{datasets}
