% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_fix_soma.R
\name{hemibrain_adjust_saved_somas}
\alias{hemibrain_adjust_saved_somas}
\title{Title heibrain_adjust_saved_somas}
\usage{
hemibrain_adjust_saved_somas(
  bodyids = NULL,
  c = NULL,
  brain = NULL,
  selected_file = "1YjkVjokXL4p4Q6BR-rGGGKWecXU370D1YMc1mgUYr8E",
  db = NULL
)
}
\arguments{
\item{bodyids}{Optional, unless using neurons method. Otherwise, list of neuron bodyids}

\item{c}{Optional. For use when correcting Cell body fibres. If not input, function will ask you to enter one later}

\item{brain}{Optional. by default will use the hemibrain surface. Other neuropil surfaces can be provided here however}

\item{selected_file}{Identifier for gsheet you wish to read from and write to. By default, Curated_splitpoints soma sheet}

\item{db}{Optional. if provided, local directory to read neurons from.}
}
\value{

}
\description{
Title heibrain_adjust_saved_somas
}
