% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain_al.surf}
\alias{hemibrain_al.surf}
\alias{hemibrain_al_microns.surf}
\title{Meshes for antennal love glomeruli}
\format{
An object of class \code{hxsurf} of length 4.

An object of class \code{hxsurf} of length 4.
}
\source{
\url{https://neuprint.janelia.org/}
}
\usage{
hemibrain_al.surf

hemibrain_al_microns.surf
}
\description{
3D meshes of 51 olfactory + 7 thermosensory antennal lobe glomeruli for the the hemibrain volume.
They are represented as a \code{\link[nat]{hxsurf}} object, in which 58 'regions' are defined,
one for each glomerulus. Meshes available both in raw voxels and in microns.
}
\details{
note that hemibrain coordinate system does not have the
  anterior-posterior axis aligned with the Z axis.
}
\examples{
\donttest{
library(nat)
\dontrun{
plot3d(hemibrain_al.surf, alpha=.3)

# can also plot single glomeruli
plot3d(subset(hemibrain_al.surf,"DM1"))

}
}
}
\seealso{
\code{\link{hemibrain_roi_meshes}}
}
\keyword{datasets}
