% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_annotations.R
\name{hemibrain_cbf}
\alias{hemibrain_cbf}
\title{Return consistent cell body fiber annotation for neurons}
\usage{
hemibrain_cbf(x, ...)
}
\arguments{
\item{x}{A vector of bodyids, a query or a data.frame compatible with
\code{\link{neuprint_ids}}}

\item{...}{Additional arguments passed to \code{\link{neuprint_get_meta}}}
}
\value{
a vector of \bold{new style} cell body fiber names.
}
\description{
Return consistent cell body fiber annotation for neurons
}
\details{
New style cell body fiber names are of the form \code{ADL01} and
match the regular expression \code{'([AP][DV][ML][0-9]+)'}.
}
\examples{
\donttest{
# these are all one lineage but some missing tract info
table(hemibrain_cbf("adPN"), useNA = 'ifany')
# Do these come from two different hemilineages?
table(hemibrain_cbf("lPN"))
}
}
\seealso{
\code{\link{neuprint_get_meta}}, \code{\link{neuprint_ids}}
}
