% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_colours.R
\docType{data}
\name{hemibrain_colours}
\alias{hemibrain_colours}
\alias{hemibrain_blender_colours}
\alias{hemibrain_blender_colors}
\alias{hemibrain_colors}
\alias{hemibrain_fig_colours}
\alias{hemibrain_fig_colors}
\alias{hemibrain_colour_ramp}
\alias{hemibrain_bright_colours}
\alias{hemibrain_bright_colors}
\alias{hemibrain_bright_colour_ramp}
\title{Some nice colours used by Schlegel and Bates et al. 2020 (in prep.) and Li et
al. 2020 (in prep)}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 6.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 22.

An object of class \code{character} of length 22.
}
\source{
\url{https://github.com/johannesbjork/LaCroixColoR}
}
\usage{
hemibrain_colours

hemibrain_blender_colours

hemibrain_blender_colors

hemibrain_colors

hemibrain_fig_colours

hemibrain_fig_colors

hemibrain_colour_ramp(n)

hemibrain_bright_colours

hemibrain_bright_colors

hemibrain_bright_colour_ramp(n)
}
\arguments{
\item{n}{the number of colours to generate}
}
\value{
a named vector of colours
}
\description{
\code{hemibrain_colours} (also \code{hemibrain_blender_colours})
  is a 6 colour palette from Li et al. 2020 (in prep). Although this has been
  described as colour blind friendly it is really not very friendly for green
  colour blind or green anomalous individuals and was actually designed to
  achieve similar perceptual brightness in Blender renderings.

  \code{hemibrain_fig_colours} is an 8 colour palette that has been used for
  some of the figures. This actually works better for colour blind
  individuals.

  \code{hemibrain_bright_colours} are nice brighter colours inspired by the
  LaCroix colour palette.
}
\examples{
\donttest{

# Li et al dark colours (Blender)
par(mar = rep(0, 4))
pie(rep(1, length(hemibrain_colours)), col = hemibrain_colours)

# Li et al figure colours
par(mar = rep(0, 4))
pie(rep(1, length(hemibrain_fig_colours)), col = hemibrain_fig_colours)

# Bright colours
par(mar = rep(0, 4))
pie(rep(1, length(hemibrain_bright_colours)), col = hemibrain_bright_colours)

}
}
\references{
\url{https://en.wikipedia.org/wiki/Color_blindness}
}
\keyword{datasets}
