% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_locality.R
\name{hemibrain_compartment_metrics}
\alias{hemibrain_compartment_metrics}
\title{Get useful metrics for each neuron's axon and dendrite}
\usage{
hemibrain_compartment_metrics(
  x,
  resample = 10,
  delta = 62.5,
  locality = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by \code{neuprintr::neuprint_read_neurons} or
possibly \code{catmaid::read.neurons.catmaid}.}

\item{resample}{stepsize to which to resample neurons. If set to \code{NULL},
neurons are not resampled.}

\item{delta}{the distance at which a synapse might occur}

\item{locality}{if TRUE \code{\link{hemibrain_overlap_locality}} is called
and the overlap locality score will be returned in addition to other
metrics.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{data.frame} of metrics broken down by putative axon and dendrute, and their overlap_locality scores
}
\description{
Calculates/retrieves useful metrics for the axon and dendrites
  of a neuron, and its split.
}
\examples{
\donttest{

# Choose neurons
## In this case some antennal lobe local neurons
al.local.neurons = c("1702323386", "2068966051", "2069311379", "1702305987", "5812996027",
"1702336197", "1793744512", "1976565858", "2007578510", "2101339904",
"5813003258", "2069647778", "1947192569", "1883788812", "1916485259",
"1887177026", "2101348562", "2132375072", "2256863785", "5813002313",
"5813054716", "5813018847", "5813055448", "1763037543", "2101391269",
"1794037618", "5813018729", "2013333009")

# Get neurons
neurons = neuprint_read_neurons(al.local.neurons)

# Convert to microns
neurons = hemibrainr:::scale_neurons(neurons)

# Split
neurons = hemibrain_flow_centrality(neurons)

# Calculate overlap and other metrics
hcm = hemibrain_compartment_metrics(neurons,
resample = NULL,
just.leaves = TRUE, # just look how much the leaves of skeletons overlap, quicker (default)
delta = 5)

# Plot
hist(hcm$overlap_locality, breaks = 100)

}
}
\seealso{
\code{\link{hemibrain_overlap_locality}}
}
