% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_reroot.R
\name{hemibrain_cut}
\alias{hemibrain_cut}
\title{Cut neurons to the hemibrain volume}
\usage{
hemibrain_cut(x, scale = 1, brain = "FAFB14", mirror = FALSE, ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} object.}

\item{scale}{scale \code{x} by a factor of \code{scale}.}

\item{brain}{the brainspace to which \code{x} is registered.}

\item{mirror}{whether or not to mirror the hemibrain volume, to which \code{x} is cut.}

\item{...}{methods supplied to \code{nat::nlapply}}
}
\value{
a \code{nat::neuronlist}.
}
\description{
Cut neurons from another data source to the hemibrain volume.
}
\examples{
\donttest{
\dontrun{
Cell07PNs.cut = hemibrain_cut(Cell07PNs.cut, brain = "FCWB")
}}
}
\seealso{
\code{\link{hemibrain_reroot}}
}
