% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_synapses.R
\name{hemibrain_extract_synapses}
\alias{hemibrain_extract_synapses}
\title{Extract synapse location and labels from a neuron/neuronlist}
\usage{
hemibrain_extract_synapses(x, prepost = c("BOTH", "PRE", "POST"), ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{prepost}{whether to get presynapses, postsynapses or both}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{data.frame}
}
\description{
Extract a single data frame describing synapse types, partners,locations and
position on a neuron's axon/dendrite.
}
\examples{
\donttest{
\dontrun{
# Choose a bodyid
id = c("452677169")

# Read in these neurons
neuron = neuprintr::neuprint_read_neurons(id)

# Re-root
neuron.flow = flow_centrality(neuron, polypre = TRUE,
mode = "centrifugal",
split = "distance")

# Let's check that this worked
syns = hemibrain_extract_synapses(neuron.flow)

# See result
plot3d_split(neuron.flow)
points3d(subset(syns,prepost==1), col = "cyan")
points3d(subset(syns,prepost==0), col = "red")

}}
}
\seealso{
\code{\link{flow_centrality}}
}
