% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_split.R
\name{hemibrain_flow_centrality}
\alias{hemibrain_flow_centrality}
\title{Split neurons into axon and dendrites using pre-computed split points}
\usage{
hemibrain_flow_centrality(
  x,
  splitpoints = hemibrainr::hemibrain_splitpoints,
  knn = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by \code{neuprintr::neuprint_read_neurons} or
possibly \code{catmaid::read.neurons.catmaid}.}

\item{splitpoints}{a \code{data.frame} of splitpoints from running \code{\link{flow_centrality}},
as produced by \code{\link{hemibrain_splitpoints}}. If a custom set of splitpoints is not given, precomputed splitpoints
are used, \code{\link{hemibrain_precomputed_splitpoints}}. This defaults to \code{hemibrain_splitpoints},
however to see the available precomputations (which have used \code{\link{flow_centrality}}, in this case `polypre = TRUE`, `mode = "centrifugal"`
and `split = "distance"`) with different parameters)
please see \code{\link{hemibrain_precomputed_splitpoints}}.}

\item{knn}{logical, whether or not to find corresponding points
between \code{splitpoints} and \code{x$d} using a nearest neighbour search in 3D
space (TRUE) or just use the given point IDs in \code{splitpoints} (i.e. if neurons
have not been resampled or their skeletons otherwise modified).}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{neuronlist}
}
\description{
Split a neuron into its putative axon and dendrite,
 using a \code{data.frame} of precomputed split points. We have already pre-computed
 splits for all neurons and stored them in this package as \code{\link{hemibrain_precomputed_splitpoints}}.
 This is helpful because  \code{\link{flow_centrality}} can take a long time to run on a large
 number of neurons.
}
\details{
From Schneider-Mizell et al. (2016): "We use flow centrality for
  four purposes. First, to split an arbor into axon and dendrite at the
  maximum centrifugal SFC, which is a preliminary step for computing the
  segregation index, for expressing all kinds of connectivity edges (e.g.
  axo-axonic, dendro-dendritic) in the wiring diagram, or for rendering the
  arbor in 3d with differently colored regions. Second, to quantitatively
  estimate the cable distance between the axon terminals and dendritic arbor
  by measuring the amount of cable with the maximum centrifugal SFC value.
  Third, to measure the cable length of the main dendritic shafts using
  centripetal SFC, which applies only to insect neurons with at least one
  output syn- apse in their dendritic arbor. And fourth, to weigh the color
  of each skeleton node in a 3d view, providing a characteristic signature of
  the arbor that enables subjective evaluation of its identity."
}
\examples{
\donttest{

# Choose some neurons
exemplars = c("202916528", "1279775082",  "203253072",
"326530038",  "203253253", "5813079341")

# Get neurons
neurons = neuprint_read_neurons(exemplars)

# Now use a pre-saved axon-dendrite split
neurons.flow = hemibrain_flow_centrality(neurons)

\dontrun{
# Plot the split to check it
nat::nopen3d()
nlscan_split(neurons.flow, WithConnectors = TRUE)
}}
}
\references{
Schneider-Mizell, C. M., Gerhard, S., Longair, M., Kazimiers, T.,
  Li, F., Zwart, M. F., … Cardona, A. (2015). Quantitative neuroanatomy for
  connectomics in Drosophila. bioRxiv, 026617. http://doi.org/10.1101/026617
}
\seealso{
\code{\link{flow_centrality}}

\code{\link{hemibrain_splitpoints}}, \code{\link{flow_centrality}}, \code{\link{hemibrain_use_splitpoints}}, \code{\link{hemibrain_precomputed_splitpoints}}
}
