% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain_metrics}
\alias{hemibrain_metrics}
\alias{hemibrain_metrics_polypre_centrifugal_distance}
\alias{hemibrain_metrics_polypre_centrifugal_synapses}
\title{Neuron metrics for all neurons in the hemibrain dataset}
\format{
An object of class \code{data.frame} with 20890 rows and 30 columns.

An object of class \code{data.frame} with 20890 rows and 30 columns.

An object of class \code{data.frame} with 20890 rows and 30 columns.
}
\source{
\url{https://elifesciences.org/articles/44590}
}
\usage{
hemibrain_metrics

hemibrain_metrics_polypre_centrifugal_distance

hemibrain_metrics_polypre_centrifugal_synapses
}
\description{
This \code{data.frame} give metrics for neurons, such as cable lengths and pre/post synapse numbers and densities,
broken down by putative compartment. E.g. axons versus dendrite. This has been calculated for neurons across the hemibrain dataset,
hemibrain:v1.0.1. The file used to generate this can be found in this package,
as \code{data-raw/hemibrain_split_all_neurons.R}. Neurons were split as described by \code{\link{hemibrain_precomputed_splitpoints}}. Note
that the first two columns, \code{pre} and \code{post} are generated by neuPrint. They differ from similar numbers in other columns of
this \code{data.frame} because 'pre' is later given as the numbers of outgoing connections (not presynapses) and some erroneous pre/post synapses will
have been removed by \code{\link{hemibrain_skeleton_check}}. Overlap locality is calculated as the 'overlap score' (Frechter et al. 2019, eLife) between
a neuron's axon and dendrite.
}
\examples{
\donttest{
head(hemibrain_metrics)
hist(hemibrain_metrics$overlap_locality,breaks = 1000)
}
}
\references{
Frechter, Shahar, Alexander Shakeel Bates, Sina Tootoonian, Michael-John Dolan, James D. Manton, Arian Rokkum Jamasb, Johannes Kohl, Davi Bock, and Gregory Sxe Jefferis. 2019.
“Functional and Anatomical Specificity in a Higher Olfactory Centre.” eLife 8 (May). https://doi.org/10.7554/eLife.44590.
}
\seealso{
\code{\link{hemibrain_flow_centrality}}, \code{\link{hemibrain_precomputed_splitpoints}}, \code{\link{axonic_cable}}, \code{\link{dendritic_cable}}
}
\keyword{datasets}
