% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_neuron_meshes.R
\name{hemibrain_neuron_meshes}
\alias{hemibrain_neuron_meshes}
\alias{hemibrain_mito_meshes}
\title{Read hemibrain neurite or mitochondrial meshes into an R neuronlist object}
\usage{
hemibrain_neuron_meshes(
  bodyids,
  lod = 2L,
  cloudvolume.url = NULL,
  conn = NULL,
  dataset = NULL,
  ...
)

hemibrain_mito_meshes(bodyids, lod = 2L, ...)
}
\arguments{
\item{bodyids}{One more hemibrain body ids (including as expressions passed
to \code{\link{neuprint_ids}})}

\item{lod}{the level of detail (highest resolution is 0, default of 2 gives a
good overall morphology while 3 is also useful and smaller still)}

\item{cloudvolume.url}{Optional a specific URL from which to download mesh
data.}

\item{conn}{A neuprint connection to the hemibrain dataset}

\item{dataset}{Optionally specify which version of the hemibrain dataset to
use (when \code{NULL} uses the most recent)}

\item{...}{Additional arguments passed to \code{\link{nlapply}}}
}
\value{
A \code{\link{neuronlist}} object
}
\description{
\code{hemibrain_neuron_meshes} reads standard neuron meshes

\code{hemibrain_mito_meshes} reads mitochondrial meshes
}
\examples{
\dontrun{
da2.lod3=hemibrain_neuron_meshes("DA2_lPN", lod=3)
da2.lod3
}
\dontrun{
da2=hemibrain_neuron_meshes("DA2_lPN")
# lod 2 ~ 5x bigger than lod 3 but does look nicer
da2
plot3d(da2)

da2mito=hemibrain_mito_meshes("DA2_lPN")

# if you just wanted one neuron (as a mesh3d object) rather than neuronlist
dl4=hemibrain_neuron_meshes("name:DL4_adPN_R")[[1]]
}
}
