% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain_olfactory_layers}
\alias{hemibrain_olfactory_layers}
\title{Hemibrain neurons in different Olfactory layers}
\format{
An object of class \code{data.frame} with 24549 rows and 25 columns.
}
\usage{
hemibrain_olfactory_layers
}
\description{
Results of an infection model used by Philipp Schlegel.
}
\keyword{datasets}
