% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain_precomputed_splitpoints}
\alias{hemibrain_precomputed_splitpoints}
\alias{hemibrain_all_splitpoints}
\alias{hemibrain_splitpoints_polypre_centrifugal_distance}
\alias{hemibrain_splitpoints_pre_centrifugal_distance}
\alias{hemibrain_splitpoints_polypre_centrifugal_synapses}
\title{Axon-dendrite split points calculated for all hemibrain neurons}
\format{
An object of class \code{data.frame} with 724603 rows and 6 columns.

An object of class \code{data.frame} with 724605 rows and 6 columns.

An object of class \code{data.frame} with 505094 rows and 6 columns.

An object of class \code{data.frame} with 724603 rows and 6 columns.
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.01.19.911453v1}
}
\usage{
hemibrain_all_splitpoints

hemibrain_splitpoints_polypre_centrifugal_distance

hemibrain_splitpoints_pre_centrifugal_distance

hemibrain_splitpoints_polypre_centrifugal_synapses
}
\description{
Axon-dendrite split points calculated for the whole hemibrain dataset,
hemibrain:v1.0.1, using a variety of parameters. The file used to generate this can be found in this
package, as \code{data-raw/hemibrain_split_all_neurons.R}. The arguments given
to \code{\link{flow_centrality}} are in the name of the data frame.
Note that, because the proportion of recovered synapses is unevenly distributed across brain
regions, after compartment detection, sometimes the dendrite was chosen as the neurite
closest to a neuron's primary branch point (data frames with 'distance' in their name), see reference.
}
\examples{
\donttest{
head(hemibrain_splitpoints_polypre_centrifugal_distance)
\dontrun{
View(hemibrain_splitpoints_polypre_centrifugal_distance)
}
}
}
\references{
Bates, A. S., P. Schlegel, R. J. V. Roberts, N. Drummond, I. F.
  M. Tamimi, R. Turnbull, X. Zhao, et al. 2020. “Complete Connectomic
  Reconstruction of Olfactory Projection Neurons in the Fly Brain.” bioRxiv.
  https://doi.org/10.1101/2020.01.19.911453.
}
\seealso{
\code{\link{hemibrain_flow_centrality}},
  \code{\link{flow_centrality}},
  \code{\link{hemibrain_use_splitpoints}}
}
\keyword{datasets}
