% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_reroot.R
\name{hemibrain_remove_bad_synapses}
\alias{hemibrain_remove_bad_synapses}
\title{Remove incorrectly placed synapses}
\usage{
hemibrain_remove_bad_synapses(
  x,
  meshes = hemibrain.surf,
  soma = TRUE,
  min.nodes.from.soma = 100,
  min.nodes.from.pnt = 5,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{meshes}{a list/a single object of class \code{mesh3d} or \code{hxsurf}.}

\item{soma}{logical, if TRUE it is assumed that the neuron being given has the soma
as its route and an intact primary neurite tract. Synapses will then be pruned from both.}

\item{min.nodes.from.soma}{the minimum number of nodes (geodesic distance) a
synapse can be from the soma. Synapses closes than this will be removed. For
comparable results across neurons, recommended to use \code{nat::resample} on
your neurons before using.}

\item{min.nodes.from.pnt}{the minimum number of nodes (geodesic distance) a
synapse can be from a point along the primary neurite.
Synapses closes than this will be removed.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{nat::neuronlist} or \code{nat::neuron} object
}
\description{
With the hemibrain projects somas and their tracts were not
separately and explicitly detected. Some false synapse assignment has happened,
adding synapses to soma and primary neurite locations, that should not exist
based on an examination of the greyscale EM data. This function removes 'out-of-mesh'
synapses, you might want to use all ROIs or the hemibrain mesh to remove these 'bad'
synapses. In addition, synapses are removed if they are too near the soma / along the
primary neurite.
}
\examples{
\donttest{
\dontrun{
# Read in a problematic neuron
neuron = neuprintr::neuprint_read_neurons(5813020793)
plot3d(neuron)
points3d(nat::xyzmatrix(neuron[[1]]$connectors))


# Re-root
neuron.fixed = hemibrain_remove_bad_synapses(neuron, meshes = hemibrain.surf)

# Let's check that this worked
plot3d(neuron.fixed)
points3d(nat::xyzmatrix(neuron.fixed[[1]]$connectors), size = 5, col = "green")

}}
}
\seealso{
\code{\link{hemibrain_reroot}}
}
