% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_reroot.R
\name{hemibrain_reroot}
\alias{hemibrain_reroot}
\title{Re-root (a) hemibrain neuron/neurons}
\usage{
hemibrain_reroot(
  x,
  method = c("manual", "estimated"),
  meshes = NULL,
  googlesheet = FALSE,
  hemibrain_somas = hemibrainr::hemibrain_somas,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{method}{whether to use the manually curated soma list or estimate soma location.}

\item{meshes}{a list/a single object of class \code{mesh3d} or \code{hxsurf}. Only used for estimation.
If \code{NULL} then \code{hemibrain_roi_meshes} is called.}

\item{googlesheet}{logical, whether to read soma locations from the \href{https://docs.google.com/spreadsheets/d/1YjkVjokXL4p4Q6BR-rGGGKWecXU370D1YMc1mgUYr8E/edit#gid=1524900531}{Google Sheet}
if \code{method == "manual"}.}

\item{hemibrain_somas}{a \code{data.frame} that gives soma locations for hemibrain neurons. See the default, \code{\link{hemibrain_somas}}. If \code{googelsheet} is \code{TRUE} this is read fresh
from the hemibrain Google team drive overseen by the Drosophila Connectomics group.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{nat::neuronlist} or \code{nat::neuron} object
}
\description{
Re-root neurons by predicting their soma location. Somas have been manually tagged by the FlyEM project.
However, in some cases the roots are wrong, or somas are outside of the volume. The Fly Connectome team at the University of Cambridge has
manually tagged better root points for these neurons. Soma locations, updated by this work, as saved as \code{\link{hemibrain_somas}} in this package.
A user can read somas from here, or get a bleeding edge version from
the \href{https://docs.google.com/spreadsheets/d/1YjkVjokXL4p4Q6BR-rGGGKWecXU370D1YMc1mgUYr8E/edit#gid=1524900531}{Google Sheet}. Alternaitvely, they
can 'estimated' a good root-point. This works by finding the longest path outside of a given neuropil mesh/meshes, as in insect
neurons the soma should be located in a cortex outside of the neuropil proper.
}
\examples{
\donttest{
# Choose some known trouble makers
bad.soma = c("5813015982","885788485","5813013913")

# Read in these neurons
neurons.bs = neuprintr::neuprint_read_neurons(bad.soma)

# Re-root
neurons = hemibrain_reroot(neurons.bs, meshes = hemibrain.surf)

\dontrun{
# Let's check that this worked
nat::nopen3d()
for(n in neurons){
     clear3d();
     message(n$bodyid)
     plot3d(n,col="brown",lwd=2)
     points3d(nat::xyzmatrix(n$connectors),col="black")
     spheres3d(nat::xyzmatrix(n)[nat::rootpoints(n),],radius=500, alpha=0.5)
     p = readline("Done? ")
}
}}
}
\seealso{
\code{\link{flow_centrality}}, \code{\link{hemibrain_somas}}
}
