% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_split.R
\name{hemibrain_split_points}
\alias{hemibrain_split_points}
\title{Get the positions of key points in a 'split' neuron}
\usage{
hemibrain_split_points(x)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by \code{neuprintr::neuprint_read_neurons} or
possibly \code{catmaid::read.neurons.catmaid}.}
}
\value{
a \code{data.frame}
}
\description{
Get the positions of key points, i.e. primary and secondary branchpoints and
a neuron's root, in a 'split' neuron.
}
\seealso{
\code{\link{flow_centrality}}
}
