% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_tags.R
\name{hemibrain_settags}
\alias{hemibrain_settags}
\alias{hemibrain_settags.neuron}
\alias{hemibrain_settags.neuronlist}
\alias{hemibrain_seetags}
\alias{hemibrain_seetags.neuron}
\alias{hemibrain_seetags.neuronlist}
\alias{hemibrain_carryover_tags}
\alias{metadata_add_tags}
\title{Set and see tags on a neuron object inc. within a neuronlist}
\usage{
hemibrain_settags(x, i = 1, ...)

\method{hemibrain_settags}{neuron}(x, i = 1, ...)

\method{hemibrain_settags}{neuronlist}(x, ...)

hemibrain_seetags(x, ...)

\method{hemibrain_seetags}{neuron}(x, ...)

\method{hemibrain_seetags}{neuronlist}(x, ...)

hemibrain_carryover_tags(x, y)

metadata_add_tags(x)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object.}

\item{i}{the index of the arguments provides through ..., to be used as the tag, if these arguments
are a vector of greater than length 1. Principally for use within method \code{hemibrainr:::list}.}

\item{...}{named arguments with values, to be added to x$tags. E.e. using soma = TRUE,
will add \code{x$tags$soma}, where \code{x$tags$soma} is \code{TRUE}. Tags are expected to be of length 1.
If x is a \code{neuronlist} then tags may be supplied of \code{length(x)}.}

\item{y}{a neuron object, which is to receive tags from x, loccated at \code{x$tags}.}
}
\description{
Set the x$tags field within a neuron object.
}
\examples{
\donttest{
# Choose neurons
## In this case some antennal lobe local neurons
al.local.neurons = c("1702323386", "2068966051")

# Get neurons
neurons = hemibrain_read_neurons(al.local.neurons)

# Tag
neurons = hemibrain_settags(neurons, antennal.lobe = TRUE)

# See tags
hemibrain_seetags(x)
}
}
\seealso{
\code{\link{hemibrain_read_neurons}}
}
