% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_meshes.R
\name{hemibrain_view}
\alias{hemibrain_view}
\title{Orient the hemibrain in an RGL viewer}
\usage{
hemibrain_view()
}
\value{
Repositions an RGL window
}
\description{
Get a good, head-on view of the hemibrain.
}
\examples{
\dontrun{
ab = neuprintr::neuprint_ROI_mesh(roi="AB(R)")
hemibrain_view()
plot3d(hemibrain.surf,col="grey",alpha=0.1)
plot3d(ab, add = TRUE, col="red",alpha=0.3)
}
}
\seealso{
\code{\link{hemibrain_skeleton_check}}, \code{\link{hemibrain.surf}},
\code{neuprintr::\link[neuprintr]{neuprint_ROI_mesh}}
}
