% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrainr-package.R
\docType{package}
\name{hemibrainr-package}
\alias{hemibrainr}
\alias{hemibrainr-package}
\title{hemibrainr: Code for Working with Data from Janelia FlyEM's Hemibrain Project``}
\description{
Useful functions for pre-processing and analysing data from Janelia FlyEM's hemibrain project <https://www.janelia.org/project-team/flyem/hemibrain>. Data can be accessed from a neuPrint instance <https://neuprint.janelia.org/> using the R package neuprintr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/hemibrainr}
  \item \url{https://flyconnectome.github.io/hemibrainr}
  \item Report bugs at \url{https://github.com/flyconnectome/hemibrainr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

}
\keyword{internal}
