% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_googledrive.R
\name{hemibrainr_googledrive_data}
\alias{hemibrainr_googledrive_data}
\alias{hemibrain_neuron_bodyids}
\alias{hemibrain_elist}
\alias{hemibrain_synapses}
\alias{hemibrain_connections}
\title{Read precomputed information from the hemibrainr Google Drive}
\usage{
hemibrain_neuron_bodyids(local = FALSE, folder = "hemibrain_neurons/")

hemibrain_elist(local = FALSE, folder = "hemibrain_neurons/")

hemibrain_synapses(local = FALSE, folder = "hemibrain_neurons/")

hemibrain_connections(local = FALSE, folder = "hemibrain_neurons/")
}
\arguments{
\item{local}{\code{FALSE} or path. By default (\code{FALSE}) data is read from \code{options()$Drive_hemibrain_data}),
but the user can specify an alternative path.}

\item{folder}{A subfolder on the hemibrain team drive or your local data folder
containing the data object to read.}
}
\value{
a \code{data.frame} or character vector
}
\description{
Read precomputed data available on the hemibrain Google Team
  Drive. This includes body IDs for all hemibrain neurons ((\code{hemibrain_neuron_bodyids})),
  all synapses (\code{hemibrain_synapses}),
  neuron-neuron connections (\code{hemibrain_connections}) and an
  edgelist  (\code{hemibrain_elist}) for all hemibrain neurons, broken down by axon and dendrite
  assignments. NBLAST matrices for all neurons against all neurons in the
  data set are also available, including ones broken down by neuron
  compartment.
}
\examples{
\donttest{
\dontrun{

# All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# For these body IDs, all synapses:
syns = hemibrain_synapses()

# For these body IDs, input and output connections:
conns = hemibrain_connections()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_elist()

}}
}
\seealso{
\code{\link{hemibrain_splitpoints}},
  \code{\link{hemibrain_flow_centrality}},
  \code{\link{hemibrainr_googledrive_data}},
  \code{\link{hemibrain_metrics}}
}
