% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_googledrive.R
\name{hemibrainr_set_drive}
\alias{hemibrainr_set_drive}
\title{Set data source for precomputed information}
\usage{
hemibrainr_set_drive(Gdrive = "hemibrain", local = "/data-raw/hemibrain_data/")
}
\arguments{
\item{Gdrive}{path to filestream mounted google drive. Can just be the name of a team
drive to be found in the standard location: \code{"/Volumes/GoogleDrive/Shared\ drives/"}.}

\item{local}{path to local folder in which data is / is to be stored.}
}
\value{
Sets options \code{Gdrive_hemibrain_data} and \code{hemibrain_data}
}
\description{
One can read precomputed data available on the hemibrain Google Team
  Drive, and other supported team drives. This includes all synapses, neuron-neuron connections and an
  edgelist for all hemibrain neurons, broken down by axon and dendrite
  assignments. NBLAST matrices for all neurons against all neurons in the
  data set are also available, including ones broken down by neuron
  compartment.
}
\examples{
\donttest{
\dontrun{

# The default drive is named 'hemibrain'
# The Wilson lab at HMS uses a drive called 'hemibrainr'.
# You could also set up your own drive with saved data.
hemibrainr_set_drive("hemibrainr")

#  All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_elist()

}}
}
\seealso{
\code{\link{hemibrain_googledrive}}
}
