% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemilineages}
\alias{hemilineages}
\title{Hemilineages namdes from T. Lee, K. Ito and V. Hartenstein}
\format{
An object of class \code{data.frame} with 123 rows and 5 columns.
}
\usage{
hemilineages
}
\description{
Hemilineages names and putative transmitter.
Transmitter guesses are based on observing at least one cell type in the literature from the hemilineage in question, known to express the given transmitter
}
\keyword{datasets}
