% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{info}
\alias{info}
\alias{pn.info}
\alias{mbon.info}
\title{Meta information of certain neuron classes}
\format{
An object of class \code{data.frame} with 333 rows and 26 columns.

An object of class \code{data.frame} with 70 rows and 30 columns.
}
\usage{
pn.info

mbon.info
}
\description{
Collated by the Flyconnectome group.
}
\keyword{datasets}
