% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_manual_split.R
\name{manually_assign_labels}
\alias{manually_assign_labels}
\alias{manually_assign_labels.neuron}
\alias{manually_assign_labels.neuronlist}
\alias{hemibrain_prune_online}
\alias{hemibrain_prune_online.neuron}
\alias{hemibrain_prune_online.neuronlist}
\alias{hemibrain_correctsoma}
\title{Manually assign the dendrite and axon to a neuron}
\usage{
manually_assign_labels(x, brain = NULL, ...)

\method{manually_assign_labels}{neuron}(x, brain = NULL, ...)

\method{manually_assign_labels}{neuronlist}(x, brain = NULL, ...)

hemibrain_prune_online(x, brain = NULL, Label = NULL, lock = NULL, ...)

\method{hemibrain_prune_online}{neuron}(x, brain = NULL, Label = NULL, lock = NULL, ...)

\method{hemibrain_prune_online}{neuronlist}(x, brain = NULL, Label = NULL, lock = NULL, ...)

hemibrain_correctsoma(x, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object}

\item{brain}{which brain to plot while splitting neuron.}

\item{...}{additional arguments passed to plotting functions, namely \code{plot3d_split}}

\item{Label}{the type of cable to be assigned/edited. Use \code{standardise_labels} to see how numbers convert to cable types.}

\item{lock}{a vector of number representing cable types to lock (so that they cannot be edited interactively be a user).
If \code{NULL} (default) no cable type is treated as locked. Locked cable will appear in green.}
}
\value{
The neuron/neuronlist object with axon/dendrite info assigned in SWC format to \code{neuron$d}
}
\description{
3D interactive functions to edit neurons in an rgl viewer,
  including re-assigned cable as being axon, dendrite, primary neurite, etc.,
  choosing soma locations and pruning bits of cable as a user chooses.
}
\examples{
\donttest{
# Choose some neurons
exemplars = c("202916528", "1279775082",  "203253072",
             "326530038",  "203253253", "5813079341")

neurons = neuprintr::neuprint_read_neurons(exemplars)

neurons.flow = hemibrain_flow_centrality(neurons)

\dontrun{
# Plot the split to check it, correcting any errors
nat::nopen3d()
neurons.manual = manually_assign_labels(neurons.flow)
}}
}
