% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{neuprint_ids}
\alias{neuprint_ids}
\title{Search neuprint for boyuds}
\usage{
neuprint_ids(x, field = "type", exact = FALSE, regex = FALSE, conn = conn, ...)
}
\arguments{
\item{x}{A set of bodyids or a query.}

\item{field}{Which field to query}

\item{exact}{whether to insist that the query string *exactly* matches the
whole of a field.}

\item{regex}{whether to treat x as a regular expression (default FALSE)}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link[neuprintr]{neuprint_login}} for details.}

\item{...}{methods passed to \code{neuprint_login}}
}
\value{
character vector of bodyids
}
\description{
Search neuprint for boyuds
}
\examples{
\donttest{
neuprint_ids("MBON")
neuprint_ids("MBON01", exact=TRUE)
neuprint_ids("MBON[0-9]+", regex=TRUE)
}
}
\seealso{
\code{\link[neuprintr]{neuprint_search}}
}
