% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{nt.cmat}
\alias{nt.cmat}
\alias{hb.nt.cmat}
\alias{fw.nt.cmat}
\title{Neurotransmitter prediction confidence matrices}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 7 rows and 7 columns.

An object of class \code{matrix} (inherits from \code{array}) with 7 rows and 7 columns.
}
\usage{
hb.nt.cmat

fw.nt.cmat
}
\description{
Neurotransmitter prediction confidence matrices for both the
hemibriain and FAFB dataset..
}
\examples{
\donttest{
head(hb.nt.cmat)
}

}
\keyword{datasets}
