% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{paper_colours}
\alias{paper_colours}
\title{Colours for labels used in Schlegel and Bates et al.2021}
\format{
An object of class \code{character} of length 178.
}
\usage{
paper_colours
}
\description{
Named vector of hex codes describing the colours used in Schlegel and Bates et al.2021 for a variety of labels.
}
\keyword{datasets}
