% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes.R
\name{plot_inspirobot}
\alias{plot_inspirobot}
\title{Plot motivations from InspiroBot}
\usage{
plot_inspirobot(cycle = NULL)
}
\arguments{
\item{cycle}{if a value, a new image will appear every \code{cycle} seconds.}
}
\value{
plots a 2D, AI generated motivational to an \code{rgl} window.
}
\description{
To help you keep going. \href{https://inspirobot.me/}{InspiroBot}
is "an artificial intelligence dedicated to generating unlimited amounts of unique
inspirational quotes for endless enrichment of pointless human existence".
}
\examples{
plot_inspirobot()
}
\references{
https://inspirobot.me/
}
\seealso{
\code{\link{hemibrain_adjust_saved_split}}
}
