% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_cable.R
\name{primary_neurite}
\alias{primary_neurite}
\alias{primary_neurite.neuron}
\alias{primary_neurite.neuprintneuron}
\alias{primary_neurite.catmaidneuron}
\alias{primary_neurite.neuronlist}
\alias{primary_branchpoint}
\alias{primary_branchpoint.neuron}
\alias{primary_branchpoint.neuronlist}
\title{Retrieve the primary neurite (cell body fibre) for a neuron}
\usage{
primary_neurite(x, neuron = TRUE, invert = FALSE, ...)

\method{primary_neurite}{neuron}(x, neuron = TRUE, invert = FALSE, ...)

\method{primary_neurite}{neuprintneuron}(x, neuron = TRUE, invert = FALSE, ...)

\method{primary_neurite}{catmaidneuron}(x, neuron = TRUE, invert = FALSE, ...)

\method{primary_neurite}{neuronlist}(x, neuron = TRUE, invert = FALSE, ...)

primary_branchpoint(x, primary_neurite = FALSE, ...)

\method{primary_branchpoint}{neuron}(x, primary_neurite = FALSE, ...)

\method{primary_branchpoint}{neuronlist}(x, primary_neurite = FALSE, ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{neuron}{logical, whether to return a neuron/neuronlist object (\code{TRUE}) or
the IDs for nodes in the primary neurite}

\item{invert}{logical, whether to return the primary neurite (\code{TRUE}) or the rest
of the neuron with its primary neurite removed (\code{FALSE}).
(note: not x$d$PointNo, but rownames(x$d)).}

\item{...}{methods sent to \code{nat::nlapply}.}

\item{primary_neurite}{logical, if \code{TRUE} only branchpoints in the cable
extracted by \code{primary_neurite} can be chosen.}
}
\value{
a \code{neuron} or \code{neuronlist} when using \code{primary neurite}.
}
\description{
Extract the cell body fibre for an insect neuron,
assuming that this neuron has a soma and that the soma is the neuron's root.
Based on the \code{nat::simplify_neuron} neuron function.
Alternatively, extract the primary branch point by a different method that
counts the numbers of leaf nodes downstream of each branch from a branch point.
The primary branchpoint is chosen as the branchpoint where
(number of leaf nodes on branch one)*(number on branch two) is the greatest.
}
\examples{
\donttest{
# Get neuron
neuron = neuprintr::neuprint_read_neurons("451987038")

# Extract primary neurite
pnt = primary_neurite(neuron)

\dontrun{
# Plot the primary neurite
nat::nopen3d()
plot3d(pnt, lwd = 2, col = "#C70E7B", soma = 1000)
plot3d(neuron, lwd = 2, col = "grey30")
}}
}
\seealso{
\code{\link{flow_centrality}}
}
