% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_cable.R
\name{prune_vertices.neuprintneuron}
\alias{prune_vertices.neuprintneuron}
\title{Prune vertices from a neuprint neuron (taking care of synapse etc information)}
\usage{
prune_vertices.neuprintneuron(x, verticestoprune, invert = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[nat]{neuron}} to prune. This can be any object that can be 
converted by \code{\link[nat]{as.ngraph}} --- see details.}

\item{verticestoprune}{An integer vector describing which vertices to remove.}

\item{invert}{Whether to keep vertices rather than dropping them (default 
FALSE).}

\item{...}{Additional arguments passed to \code{\link[nat]{as.neuron.ngraph}}}
}
\description{
Prune vertices from a neuprint neuron (taking care of synapse etc information)
}
\seealso{
\code{nat::\link[nat]{prune_vertices}}
}
