% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotes.R
\name{randomwords}
\alias{randomwords}
\title{Generate random words}
\usage{
randomwords(n = 1L, words = 1L, collapse = NULL)
}
\arguments{
\item{n}{number of random word/word combinations to generate.}

\item{words}{integer, number of words in each combination.}

\item{collapse}{Whether to combine into one charcter vector. Words
will be separated by the character entered here. Default \code{NULL}, is not to combine.}
}
\value{
returns a character vector
}
\description{
Generate random words. Wordlist from
\href{https://www.randomlists.com/random-words}{randomlists}.
}
\examples{
randomwords(10, 2, collapse = "_")
}
\references{
https://www.randomlists.com/random-words
}
\seealso{
\code{\link{plot_inspirobot}}
}
