% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_reroot.R
\name{remove_bad_synapses}
\alias{remove_bad_synapses}
\title{Remove incorrectly placed synapses}
\usage{
remove_bad_synapses(
  x,
  meshes = NULL,
  soma = TRUE,
  min.nodes.from.soma = 100,
  min.nodes.from.pnt = 5,
  primary.branchpoint = 0.25,
  method = c("unsplit", "split"),
  wipe = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by
\code{neuprintr::neuprint_read_neurons} or possibly
\code{catmaid::read.neurons.catmaid}.}

\item{meshes}{a list/a single object of class \code{mesh3d} or \code{hxsurf}.
Defaults to \code{\link{hemibrain.surf}}. See examples for alternatives.}

\item{soma}{logical, if TRUE it is assumed that the neuron being given has
the soma as its route and an intact primary neurite tract. Synapses will
then be pruned from both.}

\item{min.nodes.from.soma}{the minimum number of nodes (geodesic distance) a
synapse can be from the soma. Synapses closes than this will be removed.
For comparable results across neurons, recommended to use
\code{nat::resample} on your neurons before using.}

\item{min.nodes.from.pnt}{the minimum number of nodes (geodesic distance) a
synapse can be from a point along the primary neurite. Synapses closes than
this will be removed.}

\item{primary.branchpoint}{the proportion of the maximum branchpoint score
needed for a point to be assigned as the primary branchpoint. Used only
when \code{soma = TRUE}. See the \code{first} argument in
\code{\link{primary_neurite}} for details.}

\item{method, }{whether to estimate the primary neurite tract ('unsplit') or
use the results from flow_centrality.}

\item{wipe}{logical, whether or not to remove previous flags left on synapses
by this function.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{nat::neuronlist} or \code{nat::neuron} object
}
\description{
With the hemibrain projects somas and their tracts were not
  separately and explicitly detected. Some false synapse assignment has
  happened, adding synapses to soma and primary neurite locations, that
  should not exist based on an examination of the greyscale EM data. This
  function removes 'out-of-mesh' synapses, you might want to use all ROIs or
  the hemibrain mesh to remove these 'bad' synapses. In addition, synapses
  are removed if they are too near the soma / along the primary neurite.
}
\examples{
\donttest{
# Read in a problematic neuron
neuron = neuprint_read_neurons(5813020793)

\dontrun{
library(nat)
plot3d(neuron)
points3d(xyzmatrix(neuron[[1]]$connectors))
}

# remove any synapses outside of this surface (or on the cell body fibre)
neuron.fixed = remove_bad_synapses(neuron, meshes=hemibrain.surf)
\dontrun{
# previously we used all the individual hemibrain meshes. This may produce
# slightly different results but using *hemibrain.surf* is much faster
neuron.fixed = remove_bad_synapses(neuron, meshes=hemibrain_roi_meshes())
}
\dontrun{
library(nat)
# Let's check that this worked
plot3d(neuron.fixed)
points3d(xyzmatrix(neuron.fixed[[1]]$connectors), size = 5, col = "green")

}
\dontrun{
# Now an example with flywire neurons.
}
}
}
\seealso{
\code{\link{hemibrain_reroot}}
}
