% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_split.R
\name{add_Label}
\alias{add_Label}
\title{Manually add a Label annotation to a neuron}
\usage{
add_Label(x, Label = 2, ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by \code{neuprintr::neuprint_read_neurons} or
possibly \code{catmaid::read.neurons.catmaid}.}

\item{Label}{the Label to be added. See \code{\link{flow_centrality}}.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{neuron} or \code{neuronlist}
}
\description{
Assign all points on a skeleton with a certain Label.
}
\seealso{
\code{\link{flow_centrality}}, \code{\link{add_field}}
}
