% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_split.R
\name{add_Label}
\alias{add_Label}
\title{Manually add a Label annotation to a neuron}
\usage{
add_Label(
  x,
  PointNo = NULL,
  Label = 2,
  erase = FALSE,
  lock = NULL,
  internal.assignments = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by \code{neuprintr::neuprint_read_neurons} or
possibly \code{catmaid::read.neurons.catmaid}.}

\item{PointNo}{the points in the neuron for which the \code{Label}
will be added.If \code{NULL} all points are used.}

\item{Label}{the Label to be added. See \code{\link{flow_centrality}}.}

\item{erase}{if \code{TRUE}, all instance of \code{Label} not
in \code{PointNo} are set to \code{0}.}

\item{lock}{a numeric vector of cable types that will not be updated by \code{add_Label}.}

\item{internal.assignments}{logical. If TRUE, hidden function \code{hemibrainr::internal.assignments} is run, to assign
a neurons dendrite/axon start points, primary branch point, linker start points, etc. in line with the changes
produced by this function.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{neuron} or \code{neuronlist}
}
\description{
Manually add a Label annotation, indicative of cable type,
to a neuron's arbour and synapses.
}
\seealso{
\code{\link{flow_centrality}}, \code{\link{add_field}}
}
