% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{classed.ids}
\alias{classed.ids}
\alias{rn.ids}
\alias{orn.ids}
\alias{hrn.ids}
\alias{pn.ids}
\alias{upn.ids}
\alias{mpn.ids}
\alias{vppn.ids}
\alias{dan.ids}
\alias{mbon.ids}
\alias{alln.ids}
\alias{ton.ids}
\alias{lhn.ids}
\title{Bodyids for neuron classes}
\format{
An object of class \code{character} of length 2545.

An object of class \code{character} of length 2469.

An object of class \code{character} of length 56.

An object of class \code{character} of length 330.

An object of class \code{character} of length 137.

An object of class \code{character} of length 193.

An object of class \code{character} of length 42.

An object of class \code{character} of length 318.

An object of class \code{character} of length 67.

An object of class \code{character} of length 231.

An object of class \code{character} of length 7856.

An object of class \code{numeric} of length 2226.
}
\usage{
rn.ids

orn.ids

hrn.ids

pn.ids

upn.ids

mpn.ids

vppn.ids

dan.ids

mbon.ids

alln.ids

ton.ids

lhn.ids
}
\description{
Bodyids for neuron classes. Hard saved equivalent for \code{\link{class2ids}}.
However, you should use \code{\link{class2ids}} for the most up-to-date information.
}
\details{
neuPrint bodyids for neurons that the FlyConnectome groups at the University of Cambridge Department of Zoology,
have determined to fall into different classes.
}
\examples{
\donttest{
head(rn.ids)
}

}
\seealso{
\code{\link{class2ids}}
}
\keyword{datasets}
