% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{classed.ids}
\alias{classed.ids}
\alias{rn.ids}
\alias{orn.ids}
\alias{hrn.ids}
\alias{pn.ids}
\alias{upn.ids}
\alias{mpn.ids}
\alias{vppn.ids}
\alias{dan.ids}
\alias{mbon.ids}
\alias{alln.ids}
\alias{ton.ids}
\alias{lhn.ids}
\alias{dn.ids}
\alias{kc.ids}
\alias{apl.ids}
\alias{cent.ids}
\alias{lc.ids}
\title{Bodyids for neuron classes}
\format{
An object of class \code{character} of length 2654.

An object of class \code{character} of length 2579.

An object of class \code{character} of length 56.

An object of class \code{character} of length 338.

An object of class \code{character} of length 136.

An object of class \code{character} of length 197.

An object of class \code{character} of length 56.

An object of class \code{character} of length 318.

An object of class \code{character} of length 67.

An object of class \code{character} of length 307.

An object of class \code{character} of length 2383.

An object of class \code{numeric} of length 1428.

An object of class \code{character} of length 321.

An object of class \code{numeric} of length 2035.

An object of class \code{numeric} of length 5.

An object of class \code{numeric} of length 23.

An object of class \code{character} of length 2129.
}
\usage{
rn.ids

orn.ids

hrn.ids

pn.ids

upn.ids

mpn.ids

vppn.ids

dan.ids

mbon.ids

alln.ids

ton.ids

lhn.ids

dn.ids

kc.ids

apl.ids

cent.ids

lc.ids
}
\description{
Bodyids for neuron classes. Hard saved equivalent for \code{\link{class2ids}}.
However, you should use \code{\link{class2ids}} for the most up-to-date information.
}
\details{
neuPrint bodyids for neurons that the FlyConnectome groups at the University of Cambridge Department of Zoology,
have determined to fall into different classes.
}
\examples{
\donttest{
head(rn.ids)
}

}
\seealso{
\code{\link{class2ids}}
}
\keyword{datasets}
