% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_cable.R
\name{axonic_points}
\alias{axonic_points}
\alias{dendritic_points}
\alias{mixed_points}
\alias{primary_dendrite_points}
\alias{axonal_endings}
\alias{dendritic_endings}
\alias{axonic_endings}
\alias{primary_dendrite_endings}
\alias{axonic_cable}
\alias{dendritic_cable}
\alias{arbour_cable}
\alias{unsure_cable}
\alias{primary_dendrite_cable}
\title{Extract axonic/dendritic points/cable from a neuron/neuronlist}
\usage{
axonic_points(x, ...)

dendritic_points(x, ...)

mixed_points(x, ...)

primary_dendrite_points(x, ...)

axonal_endings(x)

dendritic_endings(x)

axonic_endings(x)

primary_dendrite_endings(x)

axonic_cable(x, mixed = FALSE, ...)

dendritic_cable(x, mixed = FALSE, ...)

arbour_cable(x, mixed = FALSE, ...)

unsure_cable(x, mixed = FALSE, ...)

primary_dendrite_cable(x, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object that has its axons/dendrites labelled in swc format in its neuron$d dataframes}

\item{...}{additional arguments passed to methods}

\item{mixed}{whether or not to include points assigned as uncertain or mixed polarity cable}
}
\value{
a matrix of 3D points
}
\description{
Extract axonic/dendritic/mixed/primary dendrite points/endpoints/cable from a neuron/neuronlist object
}
