% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_meta.R
\name{fafb_hemibrain_annotate}
\alias{fafb_hemibrain_annotate}
\alias{fafb_set_hemilineage}
\alias{fafb_set_transmitter}
\title{Set annotations for FAFB neurons in CATMAID based on hemibrain results}
\usage{
fafb_hemibrain_annotate(x, flywire = TRUE, ...)

fafb_set_hemilineage(
  find,
  ito_lee_hemilineage,
  dataset = c("FAFB", "hemibrain"),
  putative = TRUE,
  delete.find = FALSE,
  ...
)

fafb_set_transmitter(
  find,
  dataset = c("FAFB", "hemibrain"),
  putative = TRUE,
  transmitter = c("acetylcholine", "GABA", "glutamate"),
  delete.find = FALSE,
  ...
)
}
\arguments{
\item{x}{a vector of FAFB skeleton IDs (skids) that can be read with \code{catmaid::catmaid_skids()}}

\item{flywire}{logical. Whether or not to add flywire-related annotation information, e.g. a corresponding position in flywire voxel space and a matching, cognate flywire ID.}

\item{...}{arguments passed to \code{neuprintr::neuprint_get_meta} and \code{catmaid::catmaid_login}.}

\item{find}{an annotation/search term/vector of skids passed to \code{catmaid::catmaid_skids}.}

\item{ito_lee_hemilineage}{character, the correct K. Ito / T. Lee hemilineage. Must be an entry in \code{\link{hemibrain_hemilineages}}. If set to 'wipe' all lineage related annotations will be removed.}

\item{dataset}{either FAFB or hemibrain. If hemibrain, then \code{find} must be a vector of valid bodyids. Matches to FAFB neurons will be annotated for these bodyids, using \code{hemibrain_matches}.}

\item{putative}{if \code{TRUE} the word 'putative' is added to all lineage annotations to indicate to users that the labelled neurons may not have been given the correct lineages. More work may need to be done to solidify these assignments.}

\item{delete.find}{logical, is \code{TRUE} then if \code{find} is an annotation, it will be wiped from the neuron after this function is used (if you have permission to remove it)}

\item{transmitter}{character, the known or putative transmitter identity of the neurons that will be fetched using \code{find}.}
}
\value{
annotations set on CATMAID neurons in specified CATMAID instance, see your \code{catmaid::catmaid_connection()}
}
\description{
Set annotations for FAFB neurons in CATMAID based on matches made to hemibrain neurons.
E.g. transfer information on matches and cell body fibers, and also update lineage related information.
Not that \code{catmaid::flywire_matching_rewrite()} will write annotations related to flywire.
}
\examples{
\donttest{
\dontrun{

# Transfer information on the VA6 uPN
fafb_hemibrain_annotate("16")

# Annotate all GABAergic MBONs in FAFB as GABAergic
bodyids = c(1016835041,
         1048215779,
        1078693835,
         672352543,
         613719036,
         581678043,
         673085197,
         706840303,
         768555687,
         5812982924,
         424767514,
         517518166);
fafb_set_transmitter(find=bodyids,dataset="hemibrain",transmitter="GABA", putative = FALSE)

}}
}
\seealso{
\code{\link{hemibrain_get_meta}}
}
