% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywire_tracing.R
\name{flywire_annotations_for_synapses}
\alias{flywire_annotations_for_synapses}
\title{Generate a CSV of neuron synapses to import to a flywire annotation layer}
\usage{
flywire_annotations_for_synapses(
  fw.ids,
  direction = c("inputs", "outputs", "both"),
  partners = NULL,
  db = flywire_neurons(WithConnectors = TRUE),
  keep.dist.nm = 500,
  cleft_scores.thresh = 75,
  sample = 250,
  write.csv = TRUE,
  csv.path = getwd(),
  volume = NULL,
  ...
)
}
\arguments{
\item{fw.ids}{character vector, a vector of valid flywire IDs}

\item{direction}{whether to get presynapses, postsynapses or both.}

\item{partners}{character vector else not used i \code{NULL}. A vector of
valid flywire IDs for postsynaptic neurons to keep. Synapses to other
postsynaptic targets are filtered out.}

\item{db}{a \code{neuronlist} of flywire neurons with synapses attached}

\item{keep.dist.nm}{numeric, minimum distance in nm that one synapse can be
from another. One random synapse is chosen from groups of proximal
synapses. To deactivate, enter \code{NULL}.}

\item{cleft_scores.thresh}{numeric, cleft_score threshold for synapse
inclusion}

\item{sample}{numeric, the number of synapses to choose from both the
dendrite and axon of each neuron in \code{fw.ids}}

\item{write.csv}{logical, whether or not to write a \code{.csv} output file,
ready for import into flywire for each neuron. They will be named using the
pattern \code{"flywire_<fw.id>_synapse_annotations.csv"} and written to
\code{csv.path}.}

\item{csv.path}{Output path for CSV files (defaults to current working
directory)}

\item{volume}{if \code{NLLL} there is no subseting. If a \code{mesh3d}
object, only points inside this volume are chosen.}

\item{...}{further arguments passed to \code{fafbseg::flywire_partners}, when
\code{db} is \code{NULL}.}
}
\description{
Generate a CSV of neuron synapses to import to a flywire annotation layer
}
\examples{
csv = flywire_annotations_for_synapses(fw.id = "720575940616046363",
direction = "inputs",
partners = c("720575940628113292",
"720575940610903062",
"720575940611661805",
"720575940628115084",
"720575940612337575",
"720575940618976297",
"720575940619059497",
"720575940616852894",
"720575940629359076",
"720575940632437856",
"720575940624804248",
"720575940613725708"),
db = NULL, keep.dist.nm = NULL, cleft_scores.thresh = 0, sample = NULL)
}
