% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_annotations.R
\name{glomerulus}
\alias{glomerulus}
\title{Find the antennal lobe glomerulus for hemibrain body ids}
\usage{
glomerulus(bodyids, exclude.multi = FALSE)
}
\arguments{
\item{bodyids}{Vector of bodyids or a search string passed to
\code{\link{neuprint_ids}}. See details.}

\item{exclude.multi}{When \code{TRUE} multiglomerular neurons will be
returned as \code{NA}.}
}
\value{
A character vector of glomeruli named with the \code{bodyids}. If no
  glomerulus exists (or the bodyid was NA on input) \code{NA_character_} is
  returned.
}
\description{
Find the antennal lobe glomerulus for hemibrain body ids
}
\details{
Note that this is based on the same hemibrain metadata used by
  \code{\link{class2ids}}. These were curated by Marta Costa, Elizabeth
  Marin, Philipp Schlegel and Greg Jefferis in the Drosophila Connectomics
  Group / MRC LMB in Cambridge.

  \code{bodyids} may contain duplicates and \code{NA} values (in which case
  \code{NA_character_} will be returned).
}
\examples{
\dontrun{
table(glomerulus(class2ids('PN')), useNA='ifany')
table(glomerulus(class2ids('PN'), exclude.multi = TRUE), useNA='ifany')
}
}
\seealso{
\code{\link{class2ids}}, \code{\link{pn.ids}},
}
