% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain.surf}
\alias{hemibrain.surf}
\title{The hemibrain mesh}
\format{An object of class \code{hxsurf} (inherits from \code{list}) of length 4.}
\source{
\url{https://neuprint.janelia.org/}
}
\usage{
hemibrain.surf
}
\description{
A 3D mesh of the hemibrain volume as a \code{\link[nat]{hxsurf}} object.
Created by Philipp Schlegel from the union of individual ROIs in FlyEM's
hemibrain project.
}
\details{
note that hemibrain coordinate system does not have the
  anterior-posterior axis aligned with the Z axis.
}
\examples{
\donttest{
library(nat)
\dontrun{
plot3d(hemibrain.surf, alpha=.3, col='grey')

# can also convert to rgl mesh3d object for more control
hemibrain <- as.mesh3d(hemibrain.surf)
rgl::wire3d(hemibrain)
rgl::shade3d(hemibrain, alpha=.3)
rgl::shade3d(hemibrain, alpha=.3, col='grey')
}
}
}
\seealso{
\code{\link{hemibrain_roi_meshes}}
}
\keyword{datasets}
