% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_colours.R
\docType{data}
\name{hemibrain_colours}
\alias{hemibrain_colours}
\alias{hemibrain_colour_ramp}
\alias{hemibrain_bright_colours}
\alias{hemibrain_bright_colors}
\alias{hemibrain_bright_colour_ramp}
\title{Some nice colours used by Schlegel and Bates et al. 2020 (in prep.) and
Li et al. 2020 (in prep)}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 22.

An object of class \code{character} of length 22.
}
\source{
\url{https://github.com/johannesbjork/LaCroixColoR}
}
\usage{
hemibrain_colours

hemibrain_colour_ramp(n)

hemibrain_bright_colours

hemibrain_bright_colors

hemibrain_bright_colour_ramp(n)
}
\arguments{
\item{n}{the number of colours to generate}
}
\value{
a named vector of colours
}
\description{
A colour blind friendly palette from Li et al. 2020 (in prep).
Also, nicer brighter colours inspired by the LaCroix colour palette.
}
\examples{
\donttest{

# Dark colours
par(mar = rep(0, 4))
pie(rep(1, length(hemibrain_colours)), col = hemibrain_colours)

# Bright colours
par(mar = rep(0, 4))
pie(rep(1, length(hemibrain_bright_colours)), col = hemibrain_bright_colours)

}
}
\keyword{datasets}
