% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_locality.R
\name{hemibrain_compartment_metrics}
\alias{hemibrain_compartment_metrics}
\title{Get useful metrics for each neuron's axon and dendrite}
\usage{
hemibrain_compartment_metrics(
  x,
  resample = 10,
  delta = 62.5,
  locality = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by \code{neuprintr::neuprint_read_neurons} or
possibly \code{catmaid::read.neurons.catmaid}.}

\item{resample}{stepsize to which to resample neurons. If set to \code{NULL},
neurons are not resampled.}

\item{delta}{the distance at which a synapse might occur}

\item{locality}{if TRUE \code{\link{hemibrain_overlap_locality}} is called
and the overlap locality score will be returned in addition to other
metrics.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a named vector of overlap_locality scores
}
\description{
Calculates/retrieves useful metrics for the axon and dendrites
  of a neuron, and its split.
}
\seealso{
\code{\link{hemibrain_overlap_locality}}
}
