% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_synapses.R
\name{hemibrain_extract_synapses}
\alias{hemibrain_extract_synapses}
\alias{hemibrain_extract_connections}
\alias{hemibrain_extract_compartment_edgelist}
\title{Extract synapse location and labels, and edgelists from a neuron/neuronlist}
\usage{
hemibrain_extract_synapses(x, prepost = c("BOTH", "PRE", "POST"), ...)

hemibrain_extract_connections(x, prepost = c("BOTH", "PRE", "POST"), ...)

hemibrain_extract_compartment_edgelist(x, ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object. It is assumed
that this neuron has been read in by
\code{neuprintr::neuprint_read_neurons} or possibly
\code{catmaid::read.neurons.catmaid}.}

\item{prepost}{whether to get presynapses, postsynapses or both}

\item{...}{Additional arguments passed to methods or eventually to
\code{nat::\link{nlapply}}}
}
\value{

}
\description{
Extract a single data frame describing synapse/connection types,
  partners, locations and position on a neuron's axon/dendrite. You can
  either get all synapses returned or all unitary connections to a neuron's
  partners returned. Broken down by axon/dendrite (\code{Label}), and
  pre/postsynapses or pre/postsynaptic partners. Note that \code{hemibrain_extract_compartment_edgelist} will
  only return connections between neurons given in the argument \code{x}.
}
\examples{
\dontrun{
# Choose bodyids
some.pns  = sample(pn.ids, 20)

# Read in these neurons
neurons = neuprintr::neuprint_read_neurons(some.pns)

# Re-root
neurons.flow = flow_centrality(neurons,
polypre = TRUE,
mode = "centrifugal", split = "distance")

# Let's check that this worked
syns = hemibrain_extract_synapses(neurons.flow)

# Get the edgelist by compartment
elist = hemibrain_extract_compartment_edgelist(neurons.flow)

# See result
library(nat)
plot3d_split(neuron.flow)
points3d(xyzmatrix(subset(syns,prepost==1)), col = "cyan")
points3d(xyzmatrix(subset(syns,prepost==0)), col = "red")
}
}
\seealso{
\code{\link{flow_centrality}}
}
