% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_meta.R
\name{hemibrain_get_meta}
\alias{hemibrain_get_meta}
\title{Read neurons from the hemibrain connectome project}
\usage{
hemibrain_get_meta(x, ...)
}
\arguments{
\item{x}{a vector of bodyids that can be read from \url{'https://neuprint.janelia.org/'}.}

\item{...}{arguments passed to \code{neuprintr::neuprint_get_meta}.}
}
\value{
a \code{data.frame} with columns that can give a user neuron lineage and class information, as well as numbers
related to neurons' putative axon-dendrite split.
}
\description{
Read meta data for hemibrain neurons from NeuPrint and supplement this with other data available in this package.
Specifically, neuron lineage, class, olfactory system layer and FAFB match information,
as well as numbers related to neurons' putative axon-dendrite split, e.g. for synapses or cable length.
}
\examples{
\donttest{
\dontrun{

# Get neurons
meta = hemibrain_get_meta("763730414")
View(meta)

}}
}
\seealso{
\code{\link{hemibrain_splitpoints}}, \code{\link{hemibrain_flow_centrality}}, \code{\link{hemibrain_somas}},
\code{\link{hemibrain_precomputed_splitpoints}}, \code{\link{hemibrain_metrics}},\code{\link{hemibrain_olfactory_layers}}
\code{\link{hemibrain_hemilineages}}, \code{\link{classed.ids}}, \code{\link{hemibrain_read_neurons}}
}
