% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_googledrive.R
\name{hemibrain_neuron_bodyids}
\alias{hemibrain_neuron_bodyids}
\alias{hemibrain_elist}
\alias{hemibrain_synapses}
\alias{hemibrain_connections}
\title{Read precomputed information from the hemibrain Googledrive}
\usage{
hemibrain_neuron_bodyids(savedir = TRUE, local = FALSE)

hemibrain_elist(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_synapses(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_connections(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)
}
\arguments{
\item{savedir}{directory in which \code{nat::neuronlistfh} object as been saved using \code{\link{hemibrain_download_neurons}}. This acts as a database,
from which hemibrain neurons can be read. If \code{TRUE} your default save directory is used, which is stored as: \code{options()$hemibrain_data}}

\item{local}{logical, whether to try to read locally saved neurons (by default at: \code{options()$hemibrain_data}) or neurons from Google Drive (\code{options()$Gdrive_hemibrain_data}).}

\item{neuron.split}{read saved neurons spit in which way? Folder names indicative of arguments passed to \code{\link{flow_centrality}}.
and \code{\link{hemibrain_flow_centrality}}}
}
\value{
a \code{data.frame} or character vector
}
\description{
Read precomputed data available on the Hemibrain Google Team Drive. This includes all synapses, neuron-neuron connections
and an edgelist for all hemibrain neurons, broken down by axon and dendrite assigments.
}
\examples{
\donttest{
\dontrun{

#  All neurpritn IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_edgelist()

}}
}
\seealso{
\code{\link{hemibrain_splitpoints}}, \code{\link{hemibrain_flow_centrality}},
\code{\link{hemibrain_precomputed_splitpoints}}, \code{\link{hemibrain_metrics}}
}
