% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_googledrive.R
\name{hemibrain_googledrive}
\alias{hemibrain_googledrive}
\alias{hemibrain_neuron_bodyids}
\alias{hemibrain_elist}
\alias{hemibrain_synapses}
\alias{hemibrain_connections}
\alias{hemibrain_nblast}
\alias{flycircuit_neurons}
\title{Read precomputed information from the hemibrain Google Drive}
\usage{
hemibrain_neuron_bodyids(savedir = TRUE, local = FALSE)

hemibrain_elist(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_synapses(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_connections(
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

hemibrain_nblast(
  nblast = c("all", "primary.neurites", "primary.dendrites", "axons", "dendrites",
    "spines", "tracts", "arbour", "simplified"),
  savedir = TRUE,
  local = FALSE,
  neuron.split = c("polypre_centrifugal_synapses", "polypre_centrifugal_distance")
)

flycircuit_neurons(
  savedir = TRUE,
  local = FALSE,
  folder = "hemibrain_neurons/light_level/flycircuit",
  cable = c("all", "primary.neurites", "arbour"),
  data = c("neuronlist", "nblast", "dps")
)
}
\arguments{
\item{savedir}{directory in which neuron data has been deposited. If
\code{TRUE} your default save directory is used, which is stored as:
\code{getOption("Gdrive_hemibrain_data")}}

\item{local}{logical, whether to try to read locally saved neurons (by
default at: \code{getOption("hemibrain_data")}) or neurons from Google
Drive (\code{getOption("Gdrive_hemibrain_data")}).}

\item{neuron.split}{read saved neurons split in which way? Folder names
indicative of arguments passed to \code{\link{flow_centrality}}. and
\code{\link{hemibrain_flow_centrality}}}

\item{nblast}{the NBLAST matrix you would like to retrieve, e.g.
\code{"arbours"} gives you a normalised all by all NBLAST matrix of all
branching arbour.}

\item{folder}{A subfolder containing the data object to read.}

\item{cable}{the type of cable we want to read. \code{"all"} indicates full
neurons.}

\item{data}{the type of data to read, i.e. neurons, an NBLAST matrix or a
\code{\link{dotprops}} object.}
}
\value{
a \code{data.frame} or character vector
}
\description{
Read precomputed data available on the hemibrain Google Team
  Drive. This includes all synapses, neuron-neuron connections and an
  edgelist for all hemibrain neurons, broken down by axon and dendrite
  assignments. NBLAST matrices for all neurons against all neurons in the
  data set are also available, including ones broken down by neuron
  compartment.
}
\examples{
\donttest{
\dontrun{

#  All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_edgelist()

}}
}
\seealso{
\code{\link{hemibrain_splitpoints}},
  \code{\link{hemibrain_flow_centrality}},
  \code{\link{hemibrain_precomputed_splitpoints}},
  \code{\link{hemibrain_metrics}}
}
