% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_googledrive.R
\name{hemibrain_googledrive}
\alias{hemibrain_googledrive}
\alias{hemibrain_neuron_bodyids}
\alias{hemibrain_elist}
\alias{hemibrain_synapses}
\alias{hemibrain_connections}
\alias{hemibrain_nblast}
\alias{flycircuit_neurons}
\alias{hemibrain_lm_lhns}
\title{Read precomputed information from the hemibrain Google Drive}
\usage{
hemibrain_neuron_bodyids(local = FALSE, folder = "hemibrain_neurons/")

hemibrain_elist(local = FALSE, folder = "hemibrain_neurons/")

hemibrain_synapses(local = FALSE, folder = "hemibrain_neurons/")

hemibrain_connections(local = FALSE, folder = "hemibrain_neurons/")

hemibrain_nblast(
  nblast = c("hemibrain", "flywire", "hemibrain-flywire", "hemibrain-fafb14",
    "hemibrain-flycircuit", "flywire-mirror", "hemibrain-primary.neurites",
    "hemibrain-primary.dendrites", "hemibrain-axons", "hemibrain-dendrites",
    "hemibrain-spines", "hemibrain-tracts", "hemibrain-arbour", "hemibrain-simplified"),
  local = FALSE
)

flycircuit_neurons(
  local = FALSE,
  folder = "hemibrain_neurons/light_level/flycircuit",
  cable = c("all", "primary.neurites", "arbour"),
  data = c("neuronlist", "nblast", "dps"),
  brainspace = c("JRCFIB2018F", "JRCFIB2018Fraw", "FCWB")
)

hemibrain_lm_lhns(
  local = FALSE,
  folder = "hemibrain_neurons/light_level/lhns",
  data = c("neuronlist", "nblast", "dps"),
  cable = c("lhns", "lhins", "lines"),
  brainspace = c("JRCFIB2018F", "JRCFIB2018Fraw", "FCWB")
)
}
\arguments{
\item{local}{\code{FALSE} or path. By default (\code{FALSE}) data is read from \code{options()$Drive_hemibrain_data}),
but the user can specify an alternative path.}

\item{folder}{A subfolder on the hemibrain team drive or your local data folder
containing the data object to read.}

\item{nblast}{the NBLAST matrix you would like to retrieve, e.g.
\code{"arbours"} gives you a normalised all by all NBLAST matrix of all
branching arbour.}

\item{cable}{the type of cable we want to read. \code{"all"} indicates full
neurons.}

\item{data}{the type of data to read, i.e. neurons, an NBLAST matrix or a
\code{\link{dotprops}} object.}

\item{brainspace}{A template brain space for neurons loaded by
\code{hemibrain_lm_lhns}. Defaults to \code{JRCFIB2018F}.}
}
\value{
a \code{data.frame} or character vector
}
\description{
Read precomputed data available on the hemibrain Google Team
  Drive. This includes all synapses, neuron-neuron connections and an
  edgelist for all hemibrain neurons, broken down by axon and dendrite
  assignments. NBLAST matrices for all neurons against all neurons in the
  data set are also available, including ones broken down by neuron
  compartment.
}
\examples{
\donttest{
\dontrun{

#  All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_elist()

}}
}
\seealso{
\code{\link{hemibrain_splitpoints}},
  \code{\link{hemibrain_flow_centrality}},
  \code{\link{hemibrain_precomputed_splitpoints}},
  \code{\link{hemibrain_metrics}}
}
