% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain_matched}
\alias{hemibrain_matched}
\alias{lm_matched}
\title{Manually verified neuron matches between the hemibrain and FAFB dataset}
\format{
An object of class \code{data.frame} with 9154 rows and 11 columns.

An object of class \code{data.frame} with 5391 rows and 10 columns.
}
\usage{
hemibrain_matched

lm_matched
}
\description{
Manually verified neuron matches between the hemibrain and FAFB dataset, as well as some light level data from the package natverse-compatible \code{lhns}.
The Fly Connectome team at the University of Cambridge has
manually made cell-to-cell matches based on an initual NBLAST calculation, and given matches a qualitative score.
There can be no
  match (none), a poor match (poor) an okay match (okay) or an exact match (good). As a
  rule of thumb, a poor match could be a neuron from a very similar same cell
  type or a highly untraced neuron that may be the correct cell type. An okay
  match should be a neuron that looks to be from the same morphological cell
  type but there may be some discrepancies in its arbour. An exact match is a
  neuron that corresponds well between FAFB and the hemibrain data.
the \href{https://docs.google.com/spreadsheets/d/1OSlDtnR3B1LiB5cwI5x5Ql6LkZd8JOS5bBr-HTi0pOw/edit#gid=0}{Google Sheet} or using \code{\link{hemibrain_matches}}
if they have an authenticated email.
}
\keyword{datasets}
