% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_read_neurons.R
\name{hemibrain_read_neurons}
\alias{hemibrain_read_neurons}
\title{Read neurons from the hemibrain connectome project}
\usage{
hemibrain_read_neurons(x = NULL, savedir = FALSE, microns = TRUE, ...)
}
\arguments{
\item{x}{a vector of bodyids that can be read from \url{'https://neuprint.janelia.org/'}.}

\item{savedir}{directory in which \code{nat::neuronlistfh} object as been saved using \code{hemibrain_download_neurons}. This acts as a database,
from which hemibrain neurons can be read.}

\item{microns}{convert dimensions from raw voxels into microns (template brain: \code{JRCFIB2018F}, else \code{JRCFIB2018Fraw}).}

\item{...}{methods passed to \code{neuprintr::neuprint_read_neurons}, \code{\link{hemibrain_remove_bad_synapses}}
and \code{\link{hemibrain_flow_centrality}}}
}
\value{
the neuron or neuron list object inputted, with centripetal flow
  centrality information added to neuron$d and a segregation index score.
  The neuron$d$Label now gives the compartment, where axon is Label = 2,
  dendrite Label = 3, primary dendrite Label = 9 and
  primary neurite Label = 7.
  Soma is Label = 1.
}
\description{
Read neurons from the hemibrain connectome project. This function
uses the package \code{neuprintr} to read neurons. It then uses
\code{\link{hemibrain_flow_centrality}} to re-root and split neurons into putative
axons and dendrites. Optionally, it may also convert neurons from their raw voxel
space (as they are stored in neuPrint) to microns.
}
\examples{
\donttest{

# Choose neurons
## In this case some antennal lobe local neurons
al.local.neurons = c("1702323386", "2068966051", "2069311379", "1702305987", "5812996027",
"1702336197", "1793744512", "1976565858", "2007578510", "2101339904",
"5813003258", "2069647778", "1947192569", "1883788812", "1916485259",
"1887177026", "2101348562", "2132375072", "2256863785", "5813002313",
"5813054716", "5813018847", "5813055448", "1763037543", "2101391269",
"1794037618", "5813018729", "2013333009")

# Get neurons
neurons = hemibrain_read_neurons(al.local.neurons)

\dontrun{
# Plot the split to check it
nat::nopen3d()
nlscan_split(neurons, WithConnectors = TRUE)
}}
}
\seealso{
\code{\link{hemibrain_splitpoints}}, \code{\link{hemibrain_flow_centrality}},
\code{\link{hemibrain_precomputed_splitpoints}}, \code{\link{hemibrain_metrics}},\code{\link{hemibrain_remove_bad_synapses}}
}
