% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_read_neurons.R
\name{hemibrain_read_neurons}
\alias{hemibrain_read_neurons}
\alias{scale_neurons}
\alias{hemibrain_neurons}
\title{Read neurons from the hemibrain connectome project}
\usage{
hemibrain_read_neurons(
  x = NULL,
  local = FALSE,
  microns = FALSE,
  reroot = TRUE,
  remote = TRUE,
  googlesheet = FALSE,
  remove.bad.synapses = FALSE,
  clean = FALSE,
  ...
)

scale_neurons(x, scaling = (8/1000), ...)

hemibrain_neurons(
  x = NULL,
  local = FALSE,
  brain = c("JRCFIB2018Fraw", "JRCFIB2018F", "FAFB14", "JFRC2", "JRC2018F", "FCWB"),
  mirror = FALSE,
  dotprops = FALSE,
  swc = FALSE,
  zip = FALSE,
  folder = "hemibrain_neurons/"
)
}
\arguments{
\item{x}{a vector of bodyids that can be read from \url{'https://neuprint.janelia.org/'}.}

\item{local}{\code{FALSE} or path. By default (\code{FALSE}) data is read from \code{options()$Drive_hemibrain_data}), but the user can specify an alternative path.}

\item{microns}{convert dimensions from raw voxels into microns (template brain: \code{JRCFIB2018F}, else \code{JRCFIB2018Fraw}).}

\item{reroot}{logical, whether or not somas should be re-rooted.
Note that if FALSE, re-rooting occurs anyway via \code{hemibrain_flow_centrality}. However, setting this argument to \code{TRUE}}

\item{remote}{logical. Whether not to use \code{hemibrain_neurons} to try  to read neurons stored on the hemibrainr google drive, and mounted with either rclone or google filestream.}

\item{googlesheet}{logical, whether or not manually checked somas should be read from the \href{https://docs.google.com/spreadsheets/d/1YjkVjokXL4p4Q6BR-rGGGKWecXU370D1YMc1mgUYr8E/edit#gid=1524900531}{Google Sheet}}

\item{remove.bad.synapses}{whether or not to run \code{\link{hemibrain_remove_bad_synapses}} on neurons pulled from neuPrint.}

\item{clean}{whether or not to set synapse-less branches to \code{Label = 0}.}

\item{...}{arguments passed to \code{neuprintr::neuprint_read_neurons}, \code{\link{hemibrain_remove_bad_synapses}}
and \code{\link{hemibrain_flow_centrality}}}

\item{scaling}{the factor by which neuron coordinates in raw voxel space should be multiplied. The default scales to microns.}

\item{brain}{the brainspace in which hemibrain neurons have been registered. Defaults to raw voxel space for the hemibrain project.}

\item{mirror}{logical, whether or not to read neurons that have been mirrored (i.e. flipped to the 'other' brain hemisphere).}

\item{dotprops}{logical. Whether or not to retrieve a \code{nat::dotprops} object, i.e. vector cloud representations of neurons for NBLASTing.
The dotprops object will be in JRC2018FIBF micron space.}

\item{swc}{logical. When using neurons with \code{hemibrain_neurons} from the Google drive, whether to read \code{.swc} files (if \code{TRUE}), or a neuronlistfh object (default).}

\item{zip}{logical. If \code{TRUE} then \code{nat::neuronlistz} is used to read neurons from a \code{.zip} archive. Otherwise, \code{nat::neuronlistfh} is used to read neurons from a \code{.rds} file with a linked 'data' folder.
The \code{.zip} method is preferred. Both methods load a dynamic neuronlist, which only loads skeleton data into memory at the point where it is processed.}

\item{folder}{the sub-folder in which to look for a \code{nat::neuronlistfh} .rds file and its data, representing hemibrain neurons. The function will look for this
folder in the location: \code{hemibrainr:::good_savedir(local=local)},
by default the mounted Google drive (\code{options()$Gdrive_hemibrain_data}) or locally ((\code{options()$hemibrain_data}))}
}
\value{
the neuron or neuron list object inputted, with centripetal flow
  centrality information added to neuron$d and a segregation index score. The
  neuron$d$Label now gives the compartment, where axon is Label = 2, dendrite
  Label = 3, primary dendrite Label = 4 and primary neurite Label = 7. Soma
  is Label = 1.
}
\description{
Read neurons from the hemibrain connectome project. This function
uses the package \code{neuprintr} to read neurons. It then uses
\code{\link{hemibrain_flow_centrality}} to re-root and split neurons into putative
axons and dendrites. Optionally, it may also convert neurons from their raw voxel
space (as they are stored in neuPrint) to microns.
}
\examples{
\donttest{
\dontrun{

# Choose neurons
## In this case some antennal lobe local neurons
al.local.neurons = c("1702323386", "2068966051", "2069311379", "1702305987", "5812996027",
"1702336197", "1793744512", "1976565858", "2007578510", "2101339904",
"5813003258", "2069647778", "1947192569", "1883788812", "1916485259",
"1887177026", "2101348562", "2132375072", "2256863785", "5813002313",
"5813054716", "5813018847", "5813055448", "1763037543", "2101391269",
"1794037618", "5813018729", "2013333009")

# Get neurons
neurons = hemibrain_read_neurons(al.local.neurons)

# Plot the split to check it
nat::nopen3d()
nlscan_split(neurons, WithConnectors = TRUE)
}}
}
\seealso{
\code{\link{hemibrain_splitpoints}}, \code{\link{hemibrain_flow_centrality}}, \code{\link{hemibrain_somas}},
\code{\link{hemibrain_precomputed_splitpoints}}, \code{\link{hemibrain_metrics}},\code{\link{hemibrain_remove_bad_synapses}}
,\code{\link{hemibrain_get_meta}},\code{\link{flywire_neurons}}
}
