% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_reroot.R
\name{hemibrain_remove_bad_synapses}
\alias{hemibrain_remove_bad_synapses}
\title{Remove incorrectly placed synapses}
\usage{
hemibrain_remove_bad_synapses(
  x,
  meshes,
  min.nodes.from.soma = 100,
  min.nodes.from.pnt = 5,
  ...
)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{meshes}{a list/a single object of class \code{mesh3d} or \code{hxsurf}.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{nat::neuronlist} or \code{nat::neuron} object
}
\description{
With the hemibrain projects somas and their tracts were not
separately and explicitly detected. Some false synapse assignment has happened,
adding synapses to soma and primary neurite locations, that should not exist
based on an examination of the greyscale EM data. This function removes 'out-of-mesh'
synapses, you might want to use all ROIs or the hemibrain mesh to remove these 'bad'
synapses. In addition, synapses are removed if they are too near the soma / alogn the
primay neurite.
}
\examples{
\donttest{
\dontrun{
# Read in a problematic neuron
neuron = neuprintr::neuprint_read_neurons(5813020793)
plot3d(neuron)
points3d(nat::xyzmatrix(neuron[[1]]$connectors))

# Get all the roi meshes
hemibrain.rois = hemibrain_roi_meshes()

# Re-root
neuron.fixed = hemibrain_remove_bad_synapses(neuron, meshes = hemibrain.rois)

# Let's check that this worked
plot3d(neuron.fixed)
points3d(nat::xyzmatrix(neuron.fixed[[1]]$connectors), size = 5, col = "green")

}}
}
\seealso{
\code{\link{hemibrain_reroot}}
}
