% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_reroot.R
\name{hemibrain_reroot}
\alias{hemibrain_reroot}
\title{Re-root a neuron/neurons}
\usage{
hemibrain_reroot(x, meshes = hemibrain.surf, ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{meshes}{a list/a single object of class \code{mesh3d} or \code{hxsurf}.}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
a \code{nat::neuronlist} or \code{nat::neuron} object
}
\description{
Re-root neurons by predicting their soma location. This works
by finding the longest path outside of a given neuropil mesh/meshes, as in insect
neurons the soma should be located in a cortex outside of the neuropil proper.
}
\examples{
\donttest{
\dontrun{
# Choose some known troble makers
bad.soma = c("5813015982","885788485","5813013913")

# Read in these neurons
neurons.bs = neuprintr::neuprint_read_neurons(bad.soma)


# Re-root
neurons = hemibrain_reroot(neurons, meshes = hemibrain.surf)

# Let's check that this worked
nat::nopen3d()
for(n in neurons){
     clear3d();
     message(n$bodyid)
     plot3d(n,col="brown",lwd=2)
     points3d(nat::xyzmatrix(n$connectors),col="black")
     spheres3d(nat::xyzmatrix(n)[nat::rootpoints(n),],radius=500, alpha=0.5)
     p = readline("Done? ")
}
}}
}
\seealso{
\code{\link{flow_centrality}}
}
