% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_meshes.R
\name{hemibrain_roi_meshes}
\alias{hemibrain_roi_meshes}
\title{Get all the region of interest meshes in the hemibrain project}
\usage{
hemibrain_roi_meshes(microns = FALSE)
}
\arguments{
\item{microns}{convert dimensions from raw voxels into microns (template brain: \code{JRCFIB2018F}, else \code{JRCFIB2018Fraw}).}
}
\value{
a list of \code{rgl::mesh3d} objects
}
\description{
Get all the ROI meshes as named \code{rgl::mesh3d} objects.
}
\examples{
\donttest{
\dontrun{
# Get all the roi meshes
hemibrain.rois = hemibrain_roi_meshes()

# Read in these neurons
cols = rainbow(length(hemibrain.rois))
for(m in 1:length(hemibrain.rois)){
    plot3d(hemibrain.rois[[m]],alpha = 0.5, col = cols[m])
}

}}
}
\seealso{
\code{\link{hemibrain_skeleton_check}}, \code{\link{hemibrain.surf}}
}
