% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_meshes.R
\name{hemibrain_roi_meshes}
\alias{hemibrain_roi_meshes}
\title{Get some or all the region of interest meshes in the hemibrain project}
\usage{
hemibrain_roi_meshes(
  rois = NULL,
  microns = FALSE,
  superLevel = NULL,
  OmitFailures = TRUE,
  conn = NULL,
  dataset = NULL,
  ...
)
}
\arguments{
\item{rois}{The ROIs to fetch (default \code{NULL} implies all ROIs)}

\item{microns}{convert dimensions from raw voxels into microns (template brain: \code{JRCFIB2018F}, else \code{JRCFIB2018Fraw}).}

\item{superLevel}{whether not to show 'superlevel' ROIs - ROIs composed of
other ROIs. If set to NULL, both low-level and superlevel ROIs are
returned.}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in nlapply stopping with
an error message the moment there is an error. For other values, see
details.}

\item{conn}{optional, a neuprintr connection object, which also specifies the
neuPrint server. If NULL, the defaults set in your
\code{\link[=Startup]{.Rprofile}} or \code{\link[=Startup]{.Renviron}} are
used. See \code{\link[neuprintr]{neuprint_login}} for details.}

\item{dataset}{optional, a dataset you want to query. If \code{NULL}, the
default specified by your R environ file is used or, failing that the
current connection, is used. See \code{\link[neuprintr]{neuprint_login}} for details.}

\item{...}{Additional arguments passed to \code{\link{nlapply}}}
}
\value{
a \code{neuronlist} of \code{rgl::mesh3d} objects
}
\description{
This function can retrieve some or all the hemibrain project ROI
  meshes as \code{rgl::\link{mesh3d}} objects. They are returned in a named
  \code{\link{neuronlist}} object. This allows further subsets and colours to
  be applied via the standard \code{nat::\link{plot3d.neuronlist}} semantics.
  See the examples for further details.
}
\examples{
\dontrun{
# Get all the roi meshes
hemibrain.rois = hemibrain_roi_meshes()

# Read in these neurons
cols = rainbow(length(hemibrain.rois))
for(m in 1:length(hemibrain.rois)){
    plot3d(hemibrain.rois[[m]],alpha = 0.5, col = cols[m])
}

# Alternatively plot a subset all at once with rainbow colours
# making use of nat::plot3d.neuronlist
plot3d(hemibrain.rois, 1:3, add=TRUE)

# If you have neuprintr >= 1.0 you can easily find groups of meshes
# find just the terminal nodes of the MB ROI hierarchy
library(nat)
mbterms=nat::endpoints(neuprint_ROI_hierarchy(root='MB(R)', rval='graph'))
mb.rois=hemibrain_roi_meshes(mbterms)
plot3d(mb.rois, add=TRUE)
}
}
\seealso{
\code{\link{hemibrain_skeleton_check}}, \code{\link{hemibrain.surf}},
\code{neuprintr::\link[neuprintr]{neuprint_ROI_mesh}},
\code{neuprintr::\link[neuprintr]{neuprint_ROI_hierarchy}}
}
