% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemibrain_somas}
\alias{hemibrain_somas}
\title{Soma locations for hemibrain neurons}
\format{
An object of class \code{data.frame} with 22567 rows and 13 columns.
}
\usage{
hemibrain_somas
}
\description{
Somas have been manually tagged by the FlyEM project.
However, in some cases the roots are wrong, or somas are outside of the volume. The Fly Connectome team at the University of Cambridge has
manually tagged better root points for these neurons. Soma locations, updated by this work, as saved as \code{\link{hemibrain_somas}} in this package.
A user can read somas from here, or get a bleeding edge version from
the \href{https://docs.google.com/spreadsheets/d/1YjkVjokXL4p4Q6BR-rGGGKWecXU370D1YMc1mgUYr8E/edit#gid=1524900531}{Google Sheet}.
}
\keyword{datasets}
