% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_googledrive.R, R/utilities-rclone.R
\name{hemibrainr_set_drive}
\alias{hemibrainr_set_drive}
\alias{hemibrainr_team_drive}
\alias{hemibrainr_rclone}
\alias{hemibrainr_rclone_unmount}
\title{Set data source for precomputed information}
\usage{
hemibrainr_set_drive(
  Gdrive = "hemibrain",
  path = "/Volumes/GoogleDrive/Shared drives/"
)

hemibrainr_team_drive()

hemibrainr_rclone(
  Gdrive = "hemibrainr",
  path = paste0(getwd(), "hemibrainr_rclone/")
)

hemibrainr_rclone_unmount(path = paste0(getwd(), "hemibrainr_rclone/"))
}
\arguments{
\item{Gdrive}{name of the mounted Google drive. Can just be the name of a team
drive to be found in the standard location: \code{"/Volumes/GoogleDrive/Shared\ drives/"}.}

\item{path}{path to your 'hemibrainr Google drive', either a mounted remote or a local copy.
For example on a Mac, \href{https://support.google.com/drive/answer/7329379?authuser=2}{Google filestream} mounts it at: \code{/Volumes/GoogleDrive/Shared\ drives/}.
With \href{https://rclone.org/drive/}{rclone sheet} you can mount it wherever you like, by default hemibrainr expects it at:
\code{paste0(getwd(),"hemibrainr_rclone/")}}
}
\value{
Sets options \code{Gdrive_hemibrain_data}
}
\description{
One can read precomputed data available on the hemibrain Google Team
  Drive, and other supported team drives. This includes all synapses, neuron-neuron connections and an
  edgelist for all hemibrain neurons, broken down by axon and dendrite
  assignments. NBLAST matrices for all neurons against all neurons in the
  data set are also available, including ones broken down by neuron
  compartment.
}
\details{
In order to use this package to its fullest,
you need to get \code{hemibrainr} to read large amounts of data stored on a Google drive.
To do this, you have three options:

1. Mount your Google drive using \href{https://support.google.com/drive/answer/7329379?authuser=2}{Google filestream}
2. Mount your Google drive using \href{https://rclone.org/drive/}{rclone}
3. Download the Google drive and save locally

The best option is to use Google filestream.
By default, this is what \code{hemibrainr} expects.
However, you need a \href{https://workspace.google.com/pricing.html}{Google Workspace}
account (formerly G-Suite), which is a paid-for service.
Without this, the best option is to use \href{https://rclone.org/drive/}{rclone}. For detailed
instructions on how to configure rclone, please see your article in this package:
Reading hemibrainr data from Google drive.
You can also download to an external hard drive and use that.

We have two \href{https://support.google.com/a/users/answer/9310156?hl=en}{Google team drives}
 available for you to use, which contain similar data. One ("hemibrain")
 is for internal use by the \href{https://www.zoo.cam.ac.uk/research/groups/connectomics}{Drosophila Connectomics Group}.
 The other one (\code{"hemibrainr"}) is shared with those who would like access. Contact us by email to request access.
}
\examples{
\donttest{
\dontrun{
# The default drive is named 'hemibrain'
# The Wilson lab at HMS uses a drive called 'hemibrainr'.
# You could also set up your own drive with saved data.
hemibrainr_set_drive("hemibrainr")

# Ah but what it you do not have the drive loaded
# With Google filestream say, as a network drive?
# In this case, you can use the free-service rclone instead
# However, you will need to install and configure it:https://rclone.org/drive/
hemibrainr_rclone(Gdrive = "hemibrainr")
### Mounts in your working directory

# All neuprint IDs for neurons that have a split precomputed
ids = hemibrain_neuron_bodyids()

# Connectivity edgelist, broken down by axon/dendrite
elist = hemibrain_elist()

# Unmount rclone drive
hemibrainr_rclone_unmount()

}}
}
\seealso{
\code{\link{hemibrain_googledrive}}
}
