% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{hemilineages}
\alias{hemilineages}
\alias{hemibrain_hemilineages}
\title{Hemilineages names from T. Lee, K. Ito and V. Hartenstein}
\format{
An object of class \code{data.frame} with 216 rows and 8 columns.
}
\usage{
hemibrain_hemilineages
}
\description{
Hemilineages names and putative transmitter.
Transmitter guesses are based on observing at least one cell_type in the literature from the hemilineage in question, known to express the given transmitter
}
\keyword{datasets}
