% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_data.R
\docType{data}
\name{info}
\alias{info}
\alias{pn.info}
\alias{mbon.info}
\alias{ton.info}
\alias{lc.info}
\alias{alln.info}
\alias{kc.info}
\alias{cent.info}
\alias{dn.info}
\title{Meta information of certain neuron classes}
\format{
An object of class \code{data.frame} with 333 rows and 25 columns.

An object of class \code{data.frame} with 70 rows and 31 columns.

An object of class \code{data.frame} with 2383 rows and 21 columns.

An object of class \code{data.frame} with 2129 rows and 45 columns.

An object of class \code{data.frame} with 197 rows and 45 columns.

An object of class \code{data.frame} with 1927 rows and 45 columns.

An object of class \code{data.frame} with 24 rows and 45 columns.

An object of class \code{data.frame} with 431 rows and 45 columns.
}
\usage{
pn.info

mbon.info

ton.info

lc.info

alln.info

kc.info

cent.info

dn.info
}
\description{
For olfactory projection neurons (PNS), antennal lobe local neurons (ALLN),
non-MB third order olfactory neurons (TONs), lobula connection visual neurons (LC),
Kenyon cells (KC) and MB output neurons (MBONs).
Collated by the Flyconnectome group.
}
\keyword{datasets}
