% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_visualise.R
\name{plot3d_split}
\alias{plot3d_split}
\alias{nlscan_split}
\alias{plot3d_somas}
\title{Plot neurons split up by flow centrality}
\usage{
plot3d_split(
  someneuronlist,
  col = c("#1BB6AF", "#EF7C12", "#C70E7B", "#8FDA04", "#4D4D4D", "#FC6882"),
  splitnode = FALSE,
  WithConnectors = TRUE,
  WithNodes = F,
  soma = NULL,
  soma.alpha = 1,
  highflow = FALSE,
  lwd = 1,
  radius = NULL,
  already_selected = NULL,
  ...
)

nlscan_split(
  someneuronlist,
  already_selected = NULL,
  col = c("#1BB6AF", "#EF7C12", "#C70E7B", "#8FDA04", "#4D4D4D", "#FC6882"),
  WithConnectors = TRUE,
  WithNodes = FALSE,
  soma = NULL,
  radius = NULL,
  highflow = FALSE,
  Verbose = TRUE,
  Wait = TRUE,
  sleep = 0.1,
  selected_file = NULL,
  selected_col = "#fadadd",
  yaml = TRUE,
  brain = FALSE,
  ...
)

plot3d_somas(
  someneuronlist,
  col = hemibrain_bright_colors["green"],
  radius = 300,
  soma.alpha = 1,
  volume = NULL,
  invert = FALSE
)
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object that has been modified by flow.centrality.}

\item{col}{colours of sections. Defaults to orange or axons, green for primary dendrite, blue for dendrites and pink for nodes with no flow.}

\item{splitnode}{if TRUE, a magenta sphere is placed at the location of the axon-dendrite split. Possible a putative action potential initiation site?}

\item{WithConnectors}{whether to plot the anatomical location of pre (red) and post (cyan) synapses.}

\item{WithNodes}{whether to plot branch points}

\item{soma}{whether to plot a soma, and what the radius should be. If \code{NULL}, an appropriate value is guessed.}

\item{soma.alpha}{numeric, alpha transparency value for plotting the soma. Passed to \code{rgl::spheres3d}}

\item{highflow}{whether to plot the nodes of highest (with in one standard deviation less than maximum) flow centrality (pink points)}

\item{lwd}{Line width (default 1)}

\item{radius}{For connectors and axon-dendrite split node (default 1). If \code{NULL}, an appropriate value is guessed.}

\item{already_selected}{a vector of IDs for neurons that are already considered selected. If \code{NULL} then no neurons are considered selected to start with.}

\item{...}{additional arguments passed to methods.}

\item{Verbose}{logical indicating that info about each selected neuron should be printed (default TRUE)}

\item{Wait}{logical indicating that there should be a pause between each displayed neuron}

\item{sleep}{time to pause between each displayed neuron when Wait=TRUE}

\item{selected_file}{an optional path to a yaml file that already contains a selection}

\item{selected_col}{the color in which selected neurons (such as those specified in selected_file) should be plotted}

\item{yaml}{a logical indicating that selections should be saved to disk in (human-readable) yaml rather than (machine-readable) rda format}

\item{brain}{a template brain to plot. \code{FALSE} results in no brain plotted.}

\item{volume}{a \code{mesh3d} or \code{hxsurf} object. Only somas outside this volume will be plotted.}

\item{invert}{logical, if \code{TRUE} only somas outside \code{volume} will be plotted, if \code{FALSE}, only those inside.}
}
\value{
Plots coloured neuron(s)
}
\description{
Plot neurons split up by flow centrality
}
\seealso{
\code{\link{flow_centrality}}
}
