% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemibrain_cable.R
\name{primary_neurite}
\alias{primary_neurite}
\title{Retrieve the primary neurite (cell body fibre) for a neuron}
\usage{
primary_neurite(x, neuron = TRUE, invert = FALSE, ...)
}
\arguments{
\item{x}{a \code{nat::neuronlist} or \code{nat::neuron} object}

\item{neuron}{logical, whether to return a neuron/neuronlist object (TRUE) or
the IDs for nodes in the primary neurite}

\item{invert}{logical, whether to return the primary neurite (TRUE) or the rest
of the neuron with its primary neurite removed (FALSE).
(note: not x$d$PointNo, but rownames(x$d)).}

\item{...}{methods sent to \code{nat::nlapply}}
}
\value{
the neuron or neuron list object inputted, with centripetal flow
  centrality information added to neuron$d and a segregation index score.
  The neuron$d$Label now gives the compartment, where axon is Label = 2,
  dendrite Label = 3, primary dendrite Label = 9 and primary neurite Label = 7.
  Soma is Label = 1.
}
\description{
Extract the cell body fibre for an insect neuron,
assuming that this neuron has a soma and that the soma is the neuron's root.
}
\examples{
\donttest{
\dontrun{
# Get neuron
neuron = neuprintr::neuprint_read_neurons("451987038")

# Extract primary neurite
pnt = primary_neurite(neuron)

# Plot the primary neurite
nat::nopen3d()
plot3d(pnt, lwd = 2, col = "#C70E7B", soma = 1000)
plot3d(neuron, lwd = 2, col = "grey30")


}}
}
\seealso{
\code{\link{flow_centrality}}
}
